@@ -1,45 +1,123 @@
 {
   "name": "metadata-js",
   "description": "Library for building offline-first browser-based business applications",
-  "version": "0.11.223",
+  "version": "0.12.231",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
+  "engines": {
+    "node": "^8.8",
+    "npm": "^5.5"
+  },
   "directories": {
     "example": "examples",
     "doc": "doc"
   },
   "dependencies": {
-    "alasql": "^0.3",
-    "decompress": "^4.0",
-    "moment": "^2.15",
-    "node-localstorage": "^1.3",
-    "pouchdb": "^5.4",
-    "pouchdb-core": "^5.4",
-    "pouchdb-adapter-http": "^5.4",
-    "pouchdb-adapter-memory": "^5.4",
-    "pouchdb-adapter-idb": "^5.4",
-    "pouchdb-mapreduce": "^5.4",
-    "pouchdb-replication": "^5.4",
-    "pouchdb-authentication": "^0.5",
-    "rimraf": "^2.5",
-    "yargs": "^4.8"
+    "@material-ui/core": "^3.9.2",
+    "@material-ui/icons": "^3.0.2",
+    "alasql": "^0.4",
+    "cron": "^1.7.0",
+    "dataframe": "^2.0",
+    "debug": "^4.1.1",
+    "gulp-util": "^3.0.8",
+    "marked": "^0.6.1",
+    "moment": "^2.24.0",
+    "pouchdb-adapter-http": "7.0.0",
+    "pouchdb-adapter-idb": "7.0.0",
+    "pouchdb-adapter-memory": "7.0.0",
+    "pouchdb-authentication": "1.1.3",
+    "pouchdb-core": "7.0.0",
+    "pouchdb-debug": "^7.0.0",
+    "pouchdb-find": "7.0.0",
+    "pouchdb-mapreduce": "7.0.0",
+    "pouchdb-replication": "7.0.0",
+    "rc-calendar": "^9.10.10",
+    "rc-cascader": "^0.17.1",
+    "rc-time-picker": "^3.6.2",
+    "superlogin-client": "^0.8.0"
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.14",
-    "gulp": "^3.9",
-    "gulp-babel": "^6.1",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.4",
+    "@babel/plugin-external-helpers": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.3.4",
+    "@babel/plugin-proposal-decorators": "^7.3.0",
+    "@babel/plugin-proposal-do-expressions": "^7.2.0",
+    "@babel/plugin-proposal-export-default-from": "^7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-bind": "^7.2.0",
+    "@babel/plugin-proposal-function-sent": "^7.2.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-syntax-object-rest-spread": "^7.2.0",
+    "@babel/plugin-transform-object-assign": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.3.4",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "concat-files": "^0.1.1",
+    "cross-env": "^5.2.0",
+    "decompress": "^4.2",
+    "downshift": "^3.2.7",
+    "graceful-fs": "^4.1.15",
+    "gulp": "^4.0.0",
     "gulp-base64": "^0.1",
     "gulp-concat": "^2.6",
-    "gulp-csso": "^2.0",
-    "gulp-rename": "^1.2",
-    "gulp-replace": "^0.5",
-    "gulp-uglify": "^2.0",
-    "gulp-umd": "^0.2",
-    "jasmine": "^2.5",
-    "yuidocjs": "^0.10",
-    "lerna": "2.0.0-beta.28"
+    "gulp-csso": "^3.0.1",
+    "gulp-rename": "^1.4.0",
+    "gulp-replace": "^1.0.0",
+    "gulp-strip-comments": "^2.5.2",
+    "gulp-uglify": "^2.1",
+    "gulp-umd": "^2.0",
+    "gulp-wrap": "^0.14",
+    "history": "^4.7.2",
+    "md5-file": "^4.0.0",
+    "react-fullscreenable": "^2.5.1-0",
+    "react-helmet": "^5.2.0",
+    "react-onclickoutside": "^6.8.0",
+    "react-portal": "^4.2.0",
+    "react-redux": "^5.1.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-router-redux": "^5.0.0-alpha.9",
+    "react-swipeable": "^5.0.1",
+    "react-virtualized": "^9.21.0",
+    "recharts": "^1.5.0",
+    "redux": "^4.0.1",
+    "redux-dynamic-middlewares": "^1.0",
+    "redux-thunk": "^2.3.0",
+    "rimraf": "^2.6.3",
+    "rollup": "^1.6.0",
+    "rollup-plugin-cleanup": "^3.1.1",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-resolve": "^4.0.1",
+    "rollup-plugin-replace": "^2.1.0",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.29.6",
+    "yargs": "^13.2.2",
+    "yuidocjs": "^0.10"
   },
-  "engines": {
-    "node": "^6.0"
+  "peerDependecies": {
+    "jasmine": "^2.6",
+    "react": "^16.6"
+  },
+  "scripts": {
+    "build:core": "node ./packages/metadata-core/build.js",
+    "build:transition": "node ./packages/metadata-dhtmlx/build.js",
+    "build:react-ui": "cd ./packages/metadata-react/ && cross-env NODE_ENV=production BABEL_ENV=es babel --config-file ./babel.config.js ./src --out-dir . && node ./copy-files && cd ../../ && node ./packages/metadata-react/build.js",
+    "build:abstract-ui": "node ./packages/metadata-abstract-ui/build.js",
+    "build:adapter-pouchdb": "node ./packages/metadata-pouchdb/build.js",
+    "build:redux": "node ./packages/metadata-redux/build.js",
+    "build:superlogin": "node ./packages/metadata-superlogin/build.js"
   },
   "bin": {
     "metadata": "./bin/metadata-cli.js"
