@@ -1,19 +1,21 @@
 {
   "name": "metadata-core",
   "description": "The core of metadata.js as a standalone package",
-  "version": "2.0.0-beta.3",
+  "version": "2.0.18-beta.4",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
-  "dependencies": {
-    "alasql": ">=0.3",
-    "moment": ">=2.15",
-    "metadata-abstract-adapter": "^2.0.0-beta.3"
-  },
+  "license": "MIT",
   "engines": {
-    "node": ">=6.0"
+    "node": "^8.8",
+    "npm": "^6.0"
+  },
+  "dependencies": {
+    "alasql": "^0.4",
+    "debug": "^4.1",
+    "moment": "^2.24"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/oknosoft/metadata.js.git"
+    "url": "https://github.com/oknosoft/metadata.js/tree/master/packages/metadata-core"
   },
   "bugs": {
     "url": "https://github.com/oknosoft/metadata.js/issues"
@@ -23,18 +25,12 @@
   "keywords": [
     "metadata",
     "browser data engine",
+    "node data engine",
     "spa offline",
-    "rest",
-    "odata",
-    "1c",
-    "1с",
     "web сервис",
     "клиент 1с",
     "offline framework",
-    "offline data engine",
-    "CRDT",
     "offline-first",
     "replication"
-  ],
-  "license": "MIT"
+  ]
 }
