@@ -1,29 +1,37 @@
 {
   "name": "metadata-pouchdb",
   "description": "PouchDB data adapter for metadata.js",
-  "version": "2.0.0-beta.3",
+  "version": "2.0.18-beta.4",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
-  "dependencies": {
-    "metadata-abstract-adapter": "^2.0.0-beta.3",
-    "pouchdb-core": "^5.4",
-    "pouchdb-adapter-http": "^5.4",
-    "pouchdb-adapter-memory": "^5.4",
-    "pouchdb-adapter-idb": "^5.4",
-    "pouchdb-mapreduce": "^5.4",
-    "pouchdb-replication": "^5.4",
-    "pouchdb-authentication": "^0.5"
-  },
+  "license": "MIT",
   "engines": {
-    "node": ">=6.0"
+    "node": "^8.8",
+    "npm": "^6.0"
+  },
+  "dependencies": {
+    "pouchdb-adapter-http": "7.0.0",
+    "pouchdb-adapter-idb": "7.0.0",
+    "pouchdb-adapter-memory": "7.0.0",
+    "pouchdb-authentication": "1.1.3",
+    "pouchdb-core": "7.0.0",
+    "pouchdb-debug": "7.0.0",
+    "pouchdb-find": "7.0.0",
+    "pouchdb-mapreduce": "7.0.0",
+    "pouchdb-replication": "7.0.0"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/oknosoft/metadata.js.git"
+    "url": "https://github.com/oknosoft/metadata.js/tree/master/packages/metadata-pouchdb"
   },
   "bugs": {
     "url": "https://github.com/oknosoft/metadata.js/issues"
   },
   "homepage": "http://www.oknosoft.ru/metadata/",
   "main": "./index.js",
-  "license": "MIT"
+  "keywords": [
+    "metadata",
+    "couchdb",
+    "pouchdb",
+    "offline-first"
+  ]
 }
