@@ -1,15 +1,26 @@
 {
   "name": "metadata-redux",
   "description": "Actions and reducers for metadata.js",
-  "version": "2.0.0-beta.3",
+  "version": "2.0.18-beta.4",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
-  "dependencies": {},
+  "license": "MIT",
   "engines": {
-    "node": ">=6.0"
+    "node": "^8.8",
+    "npm": "^6.0"
+  },
+  "dependencies": {
+    "history": "^4.7.2",
+    "react-redux": "^5.1",
+    "react-router": "^4.3",
+    "react-router-dom": "^4.3",
+    "react-router-redux": "^5.0.0-alpha.9",
+    "redux": "^4.0.1",
+    "redux-dynamic-middlewares": "^1.0",
+    "redux-thunk": "^2.3.0"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/oknosoft/metadata.js.git"
+    "url": "https://github.com/oknosoft/metadata.js/tree/master/packages/metadata-redux"
   },
   "bugs": {
     "url": "https://github.com/oknosoft/metadata.js/issues"
@@ -23,6 +34,5 @@
     "rest",
     "odata",
     "offline-first"
-  ],
-  "license": "MIT"
+  ]
 }
