@@ -1,13 +1,26 @@
 /**
  * Конструкторы объектов данных
  *
- * &copy; Evgeniy Malyarov http://www.oknosoft.ru 2014-2016
- *
  * @module  metadata
  * @submodule meta_objs
- * @requires common
  */
 
+import utils from './utils';
+import {DataManager, DataProcessorsManager, EnumManager, RegisterManager} from './mngrs';
+import {TabularSection, TabularSectionRow} from './tabulars';
+
+class InnerData {
+
+  constructor(owner, loading) {
+    this._ts_ = {};
+    this._is_new = !(owner instanceof EnumObj);
+    this._loading = !!loading;
+    this._saving = 0;
+    this._saving_trans = false;
+    this._modified = false;
+  }
+
+}
 
 /**
  * ### Абстрактный объект данных
@@ -24,579 +37,828 @@
  * @class DataObj
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {RefDataManager}
+ * @param [loading] {Boolean}
  * @constructor
  * @menuorder 20
  * @tooltip Объект данных
  */
-class DataObj {
-
-	constructor(attr, manager) {
-
-		var tmp,
-			_ts_ = {},
-			_obj = {},
-			_data = {
-				_is_new: !(this instanceof EnumObj)
-			};
-
-		// если объект с такой ссылкой уже есть в базе, возвращаем его и не создаём нового
-		if(!(manager instanceof classes.DataProcessorsManager) && !(manager instanceof classes.EnumManager))
-			tmp = manager.get(attr, true);
-
-		if(tmp){
-			attr = null;
-			return tmp;
-		}
-
-
-		if(manager instanceof classes.EnumManager)
-			_obj.ref = attr.name;
-
-		else if(!(manager instanceof classes.RegisterManager)){
-			_obj.ref = utils.fix_guid(attr);
-
-		}else
-			_obj.ref = manager.get_ref(attr);
-
-
-		Object.defineProperties(this, {
-
-			/**
-			 * ### Фактическое хранилище данных объекта
-			 * Оно же, запись в таблице объекта локальной базы данных
-			 * @property _obj
-			 * @type Object
-			 * @final
-			 */
-			_obj: {
-				value: _obj,
-				configurable: true
-			},
-
-			/**
-			 * Хранилище ссылок на табличные части - не сохраняется в базе данных
-			 * @property _ts_
-			 */
-			_ts_: {
-				value: function( name ) {
-					if( !_ts_[name] ) {
-						_ts_[name] = new classes.TabularSection(name, this);
-					}
-					return _ts_[name];
-				},
-				configurable: true
-			},
-
-			/**
-			 * Указатель на менеджер данного объекта
-			 * @property _manager
-			 * @type DataManager
-			 * @final
-			 */
-			_manager: {
-				value : manager
-			},
-
-			/**
-			 * Пользовательские данные - аналог `AdditionalProperties` _Дополнительные cвойства_ в 1С
-			 * @property _data
-			 * @type DataManager
-			 * @final
-			 */
-			_data: {
-				value : _data,
-				configurable: true
-			}
-
-		})
-
-		if(manager.alatable && manager.push){
-			manager.alatable.push(_obj);
-			manager.push(this, _obj.ref);
-		}
-
-		attr = null;
-
-	}
-
-	_getter(f) {
-
-		var mf = this._metadata.fields[f].type,
-			res = this._obj ? this._obj[f] : "",
-			mgr, ref;
-
-		if(f == "type" && typeof res == "object")
-			return res;
-
-		else if(f == "ref"){
-			return res;
-
-		}else if(mf.is_ref){
-			if(mf.digits && typeof res === "number")
-				return res;
-
-			if(mf.hasOwnProperty("str_len") && !utils.is_guid(res))
-				return res;
-
-			if(mgr = utils.value_mgr(this._obj, f, mf)){
-				if(utils.is_data_mgr(mgr))
-					return mgr.get(res);
-				else
-					return utils.fetch_type(res, mgr);
-			}
-
-			if(res){
-				console.log([f, mf, this._obj]);
-				return null;
-			}
-
-		}else if(mf.date_part)
-			return utils.fix_date(this._obj[f], true);
-
-		else if(mf.digits)
-			return utils.fix_number(this._obj[f], !mf.hasOwnProperty("str_len"));
-
-		else if(mf.types[0]=="boolean")
-			return utils.fix_boolean(this._obj[f]);
-
-		else
-			return this._obj[f] || "";
-	}
-
-	__setter(f, v) {
-
-		var mf = this._metadata.fields[f].type,
-			mgr;
-
-		if(f == "type" && v.types)
-			this._obj[f] = v;
-
-		else if(f == "ref")
-
-			this._obj[f] = utils.fix_guid(v);
-
-		else if(mf.is_ref){
-
-			if(mf.digits && typeof v == "number" || mf.hasOwnProperty("str_len") && typeof v == "string" && !utils.is_guid(v)){
-				this._obj[f] = v;
-
-			}else {
-				this._obj[f] = utils.fix_guid(v);
-
-				mgr = utils.value_mgr(this._obj, f, mf, false, v);
-
-				if(mgr){
-					if(mgr instanceof classes.EnumManager){
-						if(typeof v == "string")
-							this._obj[f] = v;
-
-						else if(!v)
-							this._obj[f] = "";
-
-						else if(typeof v == "object")
-							this._obj[f] = v.ref || v.name || "";
-
-					}else if(v && v.presentation){
-						if(v.type && !(v instanceof DataObj))
-							delete v.type;
-						mgr.create(v);
-					}else if(!utils.is_data_mgr(mgr))
-						this._obj[f] = utils.fetch_type(v, mgr);
-				}else{
-					if(typeof v != "object")
-						this._obj[f] = v;
-				}
-			}
-
-		}else if(mf.date_part)
-			this._obj[f] = utils.fix_date(v, true);
-
-		else if(mf.digits)
-			this._obj[f] = utils.fix_number(v, !mf.hasOwnProperty("str_len"));
-
-		else if(mf.types[0]=="boolean")
-			this._obj[f] = utils.fix_boolean(v);
-
-		else
-			this._obj[f] = v;
-
-	}
-
-	__notify(f) {
-		if(!this._data._silent){
-			// TODO: observe
-			// Object.getNotifier(this).notify({
-			// 	type: 'update',
-			// 	name: f,
-			// 	oldValue: this._obj[f]
-			// });
-		}
-	}
-
-	_setter(f, v) {
-
-		if(this._obj[f] == v)
-			return;
-
-		this.__notify(f);
-		this.__setter(f, v);
-		this._data._modified = true;
-
-	}
-
-	_getter_ts(f) {
-		return this._ts_(f);
-	}
-
-	_setter_ts(f, v) {
-		var ts = this._ts_(f);
-		if(ts instanceof classes.TabularSection && Array.isArray(v))
-			ts.load(v);
-	}
-
-	/**
-	 * ### valueOf
-	 * для операций сравнения возвращаем guid
-	 */
-	valueOf(){ return this.ref }
-
-	/**
-	 * ### toJSON
-	 * для сериализации возвращаем внутренний _obj
-	 */
-	toJSON(){ return this._obj }
-
-	/**
-	 * ### toString
-	 * для строкового представления используем
-	 */
-	toString(){ return this.presentation }
-
-	/**
-	 * Метаданные текущего объекта
-	 * @property _metadata
-	 * @for DataObj
-	 * @type Object
-	 * @final
-	 */
-	get _metadata(){ return this._manager.metadata() }
-
-	/**
-	 * Пометка удаления
-	 * @property _deleted
-	 * @for DataObj
-	 * @type Boolean
-	 */
-	get _deleted(){ return !!this._obj._deleted }
-
-	/**
-	 * Признак модифицированности
-	 */
-	get _modified(){
-		if(!this._data)
-			return false;
-		return !!(this._data._modified)
-	}
-
-	/**
-	 * Возвращает "истина" для нового (еще не записанного или не прочитанного) объекта
-	 * @method is_new
-	 * @for DataObj
-	 * @return {boolean}
-	 */
-	is_new(){ return this._data._is_new }
-
-	/**
-	 * Метод для ручной установки признака _прочитан_ (не новый)
-	 */
-	_set_loaded(ref){
-		this._manager.push(this, ref);
-		this._data._modified = false;
-		this._data._is_new = false;
-	}
-
-	/**
-	 * Установить пометку удаления
-	 * @method mark_deleted
-	 * @for DataObj
-	 * @param deleted {Boolean}
-	 */
-	mark_deleted(){
-		this._obj._deleted = !!deleted;
-		this.save();
-		this.__notify('_deleted');
-	}
-
-	/**
-	 * Проверяет, является ли ссылка объекта пустой
-	 * @method empty
-	 * @return {boolean} - true, если ссылка пустая
-	 */
-	empty(){ return utils.is_empty_guid(this._obj.ref) }
-
-	/**
-	 * Читает объект из внешней или внутренней датабазы асинхронно.
-	 * В отличии от _mgr.get(), принудительно перезаполняет объект сохранёнными данными
-	 * @method load
-	 * @for DataObj
-	 * @return {Promise.<DataObj>} - промис с результатом выполнения операции
-	 * @async
-	 */
-	load() {
-
-		if (this.ref == utils.blank.guid) {
-
-			if (this instanceof CatObj)
-				this.id = "000000000";
-			else
-				this.number_doc = "000000000";
-
-			return Promise.resolve(this);
-
-		} else {
-
-			return this._manager.adapter.load_obj(this)
-				.then(() => {
-					this._data._modified = false;
-					return this;
-				});
-		}
-
-	}
-
-	/**
-	 * Освобождает память и уничтожает объект
-	 * @method unload
-	 * @for DataObj
-	 */
-	unload() {
-		var f, obj = this._obj;
-
-		this._manager.unload_obj(this.ref);
-
-		if (this._observers)
-			this._observers.length = 0;
-
-		if (this._notis)
-			this._notis.length = 0;
-
-		for (f in this._metadata.tabular_sections)
-			this[f].clear(true);
-
-		for (f in this) {
-			if (this.hasOwnProperty(f))
-				delete this[f];
-		}
-		for (f in obj)
-			delete obj[f];
-		["_ts_", "_obj", "_data"].forEach((f) => { delete this[f]; });
-		f = obj = null;
-	}
-
-	/**
-	 * ### Записывает объект
-	 * Ввыполняет подписки на события перед записью и после записи<br />
-	 * В зависимости от настроек, выполняет запись объекта во внешнюю базу данных
-	 *
-	 * @method save
-	 * @for DataObj
-	 * @param [post] {Boolean|undefined} - проведение или отмена проведения или просто запись
-	 * @param [operational] {Boolean} - режим проведения документа (Оперативный, Неоперативный)
-	 * @param [attachments] {Array} - массив вложений
-	 * @return {Promise.<DataObj>} - промис с результатом выполнения операции
-	 * @async
-	 */
-	save(post, operational, attachments) {
-
-		if (this instanceof DocObj && typeof post == "boolean") {
-			var initial_posted = this.posted;
-			this.posted = post;
-		}
-
-		var saver,
-
-			before_save_res = this._manager.handle_event(this, "before_save"),
-
-			reset_modified = function () {
-
-				if (before_save_res === false) {
-					if (this instanceof DocObj && typeof initial_posted == "boolean" && this.posted != initial_posted) {
-						this.posted = initial_posted;
-					}
-				} else
-					this._data._modified = false;
-
-				saver = null;
-				before_save_res = null;
-				reset_modified = null;
-
-				return this;
-			}.bind(this);
-
-		// если процедуры перед записью завершились неудачно или запись выполнена нестандартным способом - не продолжаем
-		if (before_save_res === false) {
-			return Promise.reject(reset_modified());
-
-		} else if (before_save_res instanceof Promise || typeof before_save_res === "object" && before_save_res.then) {
-			// если пользовательский обработчик перед записью вернул промис, его и возвращаем
-			return before_save_res.then(reset_modified);
-		}
-
-
-		// для объектов с иерархией установим пустого родителя, если иной не указан
-		if (this._metadata.hierarchical && !this._obj.parent)
-			this._obj.parent = utils.blank.guid;
-
-		// для документов, контролируем заполненность даты
-		if (this instanceof DocObj || this instanceof TaskObj || this instanceof BusinessProcessObj) {
-
-			if (utils.blank.date == this.date)
-				this.date = new Date();
-
-			if (!this.number_doc)
-				this.new_number_doc();
-
-		} else {
-			if (!this.id)
-				this.new_number_doc();
-		}
-
-
-		// если не указаны обязательные реквизиты
-		// TODO: show_msg alert-error
-		// if (msg && msg.show_msg) {
-		// 	for (var mf in this._metadata.fields) {
-		// 		if (this._metadata.fields[mf].mandatory && !this._obj[mf]) {
-		// 			msg.show_msg({
-		// 				title: msg.mandatory_title,
-		// 				type: "alert-error",
-		// 				text: msg.mandatory_field.replace("%1", this._metadata.fields[mf].synonym)
-		// 			});
-		// 			before_save_res = false;
-		// 			return Promise.reject(reset_modified());
-		// 		}
-		// 	}
-		// }
-
-		// в зависимости от типа кеширования, получаем saver и сохраняем объект во внешней базе
-		return this._manager.adapter.save_obj(
-			this, {
-				post: post,
-				operational: operational,
-				attachments: attachments
-			})
-		// и выполняем обработку после записи
-			.then(function (obj) {
-				return obj._manager.handle_event(obj, "after_save");
-			})
-			.then(reset_modified);
-	}
-
-
-	/**
-	 * ### Возвращает присоединенный объект или файл
-	 * @method get_attachment
-	 * @for DataObj
-	 * @param att_id {String} - идентификатор (имя) вложения
-	 */
-	get_attachment(att_id) {
-		return this._manager.get_attachment(this.ref, att_id);
-	}
-
-	/**
-	 * ### Сохраняет объект или файл во вложении
-	 * Вызывает {{#crossLink "DataManager/save_attachment:method"}} одноименный метод менеджера {{/crossLink}} и передаёт ссылку на себя в качестве контекста
-	 *
-	 * @method save_attachment
-	 * @for DataObj
-	 * @param att_id {String} - идентификатор (имя) вложения
-	 * @param attachment {Blob|String} - вложениe
-	 * @param [type] {String} - mime тип
-	 * @return Promise.<DataObj>
-	 * @async
-	 */
-	save_attachment(att_id, attachment, type) {
-		return this._manager.save_attachment(this.ref, att_id, attachment, type)
-			.then(function (att) {
-				if (!this._attachments)
-					this._attachments = {};
-				if (!this._attachments[att_id] || !att.stub)
-					this._attachments[att_id] = att;
-				return att;
-			}.bind(this));
-	}
-
-
-	/**
-	 * ### Удаляет присоединенный объект или файл
-	 * Вызывает одноименный метод менеджера и передаёт ссылку на себя в качестве контекста
-	 *
-	 * @method delete_attachment
-	 * @for DataObj
-	 * @param att_id {String} - идентификатор (имя) вложения
-	 * @async
-	 */
-	delete_attachment(att_id) {
-		return this._manager.delete_attachment(this.ref, att_id)
-			.then(function (att) {
-				if (this._attachments)
-					delete this._attachments[att_id];
-				return att;
-			}.bind(this));
-	}
-
-
-	/**
-	 * ### Включает тихий режим
-	 * Режим, при котором объект не информирует мир об изменениях своих свойств.<br />
-	 * Полезно, например, при групповых изменениях, чтобы следящие за объектом формы не тратили время на перерисовку при изменении каждого совйтсва
-	 *
-	 * @method _silent
-	 * @for DataObj
-	 * @param [v] {Boolean}
-	 */
-	_silent(v) {
-		if (typeof v == "boolean")
-			this._data._silent = v;
-		else {
-			this._data._silent = true;
-			setTimeout(function () {
-				this._data._silent = false;
-			}.bind(this));
-		}
-	}
-
-	_mixin_attr(attr){
-
-		if(attr && typeof attr == "object"){
-			if(attr._not_set_loaded){
-				delete attr._not_set_loaded;
-				utils._mixin(this, attr);
-
-			}else{
-				utils._mixin(this, attr);
-
-				if(!utils.is_empty_guid(this.ref) && (attr.id || attr.name))
-					this._set_loaded(this.ref);
-			}
-		}
-	}
-
-
-	/**
-	 * ### Выполняет команду печати
-	 * Вызывает одноименный метод менеджера и передаёт себя в качестве объекта печати
-	 *
-	 * @method print
-	 * @for DataObj
-	 * @param model {String} - идентификатор макета печатной формы
-	 * @param [wnd] - указатель на форму, из которой произведён вызов команды печати
-	 * @return {*|{value}|void}
-	 * @async
-	 */
-	print(model, wnd) {
-		return this._manager.print(this, model, wnd);
-	}
+export class DataObj {
+
+  constructor(attr, manager, loading, direct) {
+
+    // если объект с такой ссылкой уже есть в базе, возвращаем его и не создаём нового
+    if(!(manager instanceof DataProcessorsManager) && !(manager instanceof EnumManager)) {
+      const tmp = manager.get(attr, true);
+      if(tmp) {
+        return tmp;
+      }
+    }
+
+    Object.defineProperties(this, {
+
+      /**
+       * ### Фактическое хранилище данных объекта
+       * Оно же, запись в таблице объекта локальной базы данных
+       * @property _obj
+       * @type Object
+       * @final
+       */
+      _obj: {
+        value: direct ? attr : {
+          ref: manager instanceof EnumManager ? attr.name : (manager instanceof RegisterManager ? manager.get_ref(attr) : utils.fix_guid(attr))
+        },
+        configurable: true
+      },
+
+      /**
+       * Указатель на менеджер данного объекта
+       * @property _manager
+       * @type DataManager
+       * @final
+       */
+      _manager: {
+        value: manager
+      },
+
+      /**
+       * Внутренние и пользовательские данные - аналог `AdditionalProperties` _Дополнительные cвойства_ в 1С
+       * @property _data
+       * @type InnerData
+       * @final
+       */
+      _data: {
+        value: new InnerData(this, loading),
+        configurable: true
+      }
+
+    });
+
+    if(manager.alatable && manager.push) {
+      manager.alatable.push(this._obj);
+      manager.push(this, this._obj.ref);
+    }
+
+  }
+
+  _getter(f) {
+
+    const mf = this._metadata(f).type;
+    const {_obj} = this;
+    const res = _obj ? _obj[f] : '';
+
+    if(f == 'type' && typeof res == 'object') {
+      return res;
+    }
+    else if(f == 'ref') {
+      return res;
+    }
+    else if(mf.is_ref) {
+
+      if(mf.digits && typeof res === 'number') {
+        return res;
+      }
+
+      if(mf.hasOwnProperty('str_len') && !utils.is_guid(res)) {
+        return res;
+      }
+
+      let mgr = this._manager.value_mgr(_obj, f, mf);
+      if(mgr) {
+        if(utils.is_data_mgr(mgr)) {
+          return mgr.get(res, false, false);
+        }
+        else {
+          return utils.fetch_type(res, mgr);
+        }
+      }
+
+      if(res) {
+        console.log([f, mf, _obj]);
+        return null;
+      }
+
+    }
+    else if(mf.date_part) {
+      return utils.fix_date(_obj[f], true);
+    }
+    else if(mf.digits) {
+      return utils.fix_number(_obj[f], !mf.hasOwnProperty('str_len'));
+    }
+    else if(mf.types[0] == 'boolean') {
+      return utils.fix_boolean(_obj[f]);
+    }
+    else {
+      return _obj[f] || '';
+    }
+  }
+
+  /**
+   * Устанваливает значение реквизита с приведением типов без проверки, отличается ли оно от предыдущего
+   * @param f
+   * @param v
+   * @param [mf]
+   * @private
+   */
+  __setter(f, v, mf) {
+
+    const {_obj, _data} = this;
+
+    if(!mf){
+      mf = this._metadata(f).type;
+    }
+
+    // выполняем value_change с блокировкой эскалации
+    if(!_data._loading) {
+      _data._loading = true;
+      const res = this.value_change(f, mf, v);
+      _data._loading = false;
+      if(res === false) {
+        return;
+      }
+    }
+
+    if(f === 'type' && v.types) {
+      _obj[f] = v;
+    }
+    else if(f === 'ref') {
+      _obj[f] = utils.fix_guid(v);
+    }
+    else if(mf instanceof DataObj || mf instanceof DataManager) {
+      _obj[f] = utils.fix_guid(v, false);
+    }
+    else if(mf.is_ref) {
+
+      if(mf.digits && typeof v === 'number' || mf.hasOwnProperty('str_len') && typeof v === 'string' && !utils.is_guid(v)) {
+        _obj[f] = v;
+      }
+      else if(typeof v === 'boolean' && mf.types.indexOf('boolean') != -1) {
+        _obj[f] = v;
+      }
+      else if(mf.date_part && v instanceof Date) {
+        _obj[f] = v;
+      }
+      else {
+        _obj[f] = utils.fix_guid(v);
+
+        if(utils.is_data_obj(v) && mf.types.indexOf(v._manager.class_name) != -1) {
+
+        }
+        else {
+          let mgr = this._manager.value_mgr(_obj, f, mf, false, v);
+          if(mgr) {
+            if(mgr instanceof EnumManager) {
+              if(typeof v === 'string') {
+                _obj[f] = v;
+              }
+              else if(!v) {
+                _obj[f] = '';
+              }
+              else if(typeof v === 'object') {
+                _obj[f] = v.ref || v.name || '';
+              }
+
+            }
+            else if(v && v.presentation) {
+              if(v.type && !(v instanceof DataObj)) {
+                delete v.type;
+              }
+              mgr.create(v);
+            }
+            else if(!utils.is_data_mgr(mgr)) {
+              _obj[f] = utils.fetch_type(v, mgr);
+            }
+          }
+          else {
+            if(typeof v !== 'object') {
+              _obj[f] = v;
+            }
+          }
+        }
+      }
+    }
+    else if(mf.date_part) {
+      _obj[f] = utils.fix_date(v, !mf.hasOwnProperty('str_len'));
+    }
+    else if(mf.digits) {
+      _obj[f] = utils.fix_number(v, !mf.hasOwnProperty('str_len'));
+    }
+    else if(mf.types[0] == 'boolean') {
+      _obj[f] = utils.fix_boolean(v);
+    }
+    else {
+      _obj[f] = v;
+    }
+
+  }
+
+  __notify(f) {
+    const {_data, _manager} = this;
+    if(_data && !_data._loading) {
+      _data._modified = true;
+      _manager.emit_async('update', this, {[f]: this._obj[f]});
+    }
+  }
+
+  /**
+   * Устанваливает значение, если оно отличается от предыдущего
+   * @param f
+   * @param v
+   * @private
+   */
+  _setter(f, v) {
+    if(this._obj[f] != v) {
+      this.__notify(f);
+      this.__setter(f, v);
+    }
+  }
+
+  /**
+   * Получает (при необходимости - конструирует) табличную часть
+   * @param f {String} - имя табчасти
+   * @return {TabularSection}
+   * @private
+   */
+  _getter_ts(f) {
+    const {_ts_} = this._data;
+    return _ts_[f] || (_ts_[f] = new TabularSection(f, this));
+  }
+
+  _setter_ts(f, v) {
+    const ts = this._getter_ts(f);
+    ts instanceof TabularSection && Array.isArray(v) && ts.load(v);
+  }
+
+  /**
+   * ### valueOf
+   * для операций сравнения возвращаем guid
+   */
+  valueOf() {
+    return this.ref;
+  }
+
+  /**
+   * ### toJSON
+   * для сериализации возвращаем внутренний _obj
+   */
+  toJSON() {
+    return this._obj;
+  }
+
+  /**
+   * ### toString
+   * для строкового представления используем
+   */
+  toString() {
+    return this.presentation;
+  }
+
+  /**
+   * Метаданные текущего объекта
+   * @method _metadata
+   * @for DataObj
+   * @param field_name
+   * @type Object
+   * @final
+   */
+  _metadata(field_name) {
+    return this._manager.metadata(field_name);
+  }
+
+  /**
+   * Пометка удаления
+   * @property _deleted
+   * @for DataObj
+   * @type Boolean
+   */
+  get _deleted() {
+    return !!this._obj._deleted;
+  }
+  set _deleted(v) {
+    this._obj._deleted = !!v;
+  }
+
+  /**
+   * ### Ревизия
+   * Eё устанваливает адаптер при чтении и записи
+   * @property _rev
+   * @for DataObj
+   * @type Boolean
+   */
+  get _rev() {
+    return this._obj._rev || '';
+  }
+  set _rev(v) {
+  }
+
+  /**
+   * Признак модифицированности
+   */
+  get _modified() {
+    return !!this._data._modified;
+  }
+  set _modified(v) {
+    this._data._modified = !!v;
+  }
+
+  /**
+   * Возвращает "истина" для нового (еще не записанного или не прочитанного) объекта
+   * @method is_new
+   * @for DataObj
+   * @return {boolean}
+   */
+  is_new() {
+    return !this._data || this._data._is_new;
+  }
+
+  /**
+   * Метод для ручной установки признака _прочитан_ (не новый)
+   */
+  _set_loaded(ref) {
+    this._manager.push(this, ref);
+    Object.assign(this._data, {
+      _modified: false,
+      _is_new: false,
+      _loading: false,
+    });
+    return this;
+  }
+
+  /**
+   * Установить пометку удаления
+   * @method mark_deleted
+   * @for DataObj
+   * @param deleted {Boolean}
+   */
+  mark_deleted(deleted) {
+    this._obj._deleted = !!deleted;
+    return this.save();
+  }
+
+  get class_name() {
+    return this._manager.class_name;
+  }
+  set class_name(v) {
+    return this._obj.class_name = v;
+  }
+
+  /**
+   * Проверяет, является ли ссылка объекта пустой
+   * @method empty
+   * @return {boolean} - true, если ссылка пустая
+   */
+  empty() {
+    return !this._obj || utils.is_empty_guid(this._obj.ref);
+  }
+
+  /**
+   * Читает объект из внешней или внутренней датабазы асинхронно.
+   * В отличии от _mgr.get(), принудительно перезаполняет объект сохранёнными данными
+   * @method load
+   * @for DataObj
+   * @return {Promise.<DataObj>} - промис с результатом выполнения операции
+   * @async
+   */
+  load() {
+    const {_data} = this;
+    if(this.ref == utils.blank.guid) {
+      if(_data) {
+        _data._loading = false;
+        _data._modified = false;
+      }
+      return Promise.resolve(this);
+    }
+    else if(_data._loading) {
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+          resolve(_data._loading ? this.load() : this);
+        }, 1000);
+      });
+    }
+    else {
+      _data._loading = true;
+      return this._manager.adapter.load_obj(this)
+        .then(() => {
+          _data._loading = false;
+          _data._modified = false;
+          return this.after_load();
+        });
+    }
+  }
+
+  /**
+   * Освобождает память и уничтожает объект
+   * @method unload
+   * @for DataObj
+   */
+  unload() {
+    const {_obj, ref, _data, _manager} = this;
+    _manager.unload_obj(ref);
+    _data._loading = true;
+    //_manager.emit_async('unload', this);
+    for (const ts in this._metadata().tabular_sections) {
+      this[ts].clear();
+    }
+    for (const f in this) {
+      if(this.hasOwnProperty(f)) {
+        delete this[f];
+      }
+    }
+    for (const f in _obj) {
+      delete _obj[f];
+    }
+    delete this._obj;
+  }
+
+  /**
+   * ### Записывает объект
+   * Ввыполняет подписки на события перед записью и после записи<br />
+   * В зависимости от настроек, выполняет запись объекта во внешнюю базу данных
+   *
+   * @method save
+   * @for DataObj
+   * @param [post] {Boolean|undefined} - проведение или отмена проведения или просто запись
+   * @param [operational] {Boolean} - режим проведения документа (Оперативный, Неоперативный)
+   * @param [attachments] {Array} - массив вложений
+   * @return {Promise.<DataObj>} - промис с результатом выполнения операции
+   * @async
+   */
+  save(post, operational, attachments) {
+
+    if(utils.is_empty_guid(this.ref)) {
+      return Promise.resolve(this);
+    }
+
+    // запоминаем признак проведенности, чтобы восстановить его в случае неудачной записи
+    let initial_posted;
+    if(this instanceof DocObj && typeof post == 'boolean') {
+      initial_posted = this.posted;
+      this.posted = post;
+    }
+
+    // выполняем обработчик перед записью
+    const {_data} = this;
+    _data._saving_trans = true;
+    return Promise.resolve()
+      .then(() => this.before_save())
+      .then((before_save_res) => {
+
+        // этот код выполним в самом конце, после записи
+        const reset_modified = () => {
+          if(before_save_res === false) {
+            if(this instanceof DocObj && typeof initial_posted == 'boolean' && this.posted !== initial_posted) {
+              this.posted = initial_posted;
+            }
+          }
+          else {
+            _data._modified = false;
+          }
+          _data._saving = 0;
+          _data._saving_trans = false;
+          return this;
+        };
+
+
+        // если процедуры перед записью завершились неудачно или запись выполнена нестандартным способом - не продолжаем
+        if(before_save_res === false) {
+          return Promise.reject(reset_modified());
+        }
+
+        // этот код выполняем в случае ошибки незаполненных реквизитов
+        const reset_mandatory = (msg) => {
+          before_save_res = false;
+          reset_modified();
+          md.emit('alert', msg);
+          const err = new Error(msg.text);
+          err.msg = msg;
+          return Promise.reject(err);
+        };
+
+        // для объектов с иерархией установим пустого родителя, если иной не указан
+        if(this._metadata().hierarchical && !this._obj.parent) {
+          this._obj.parent = utils.blank.guid;
+        }
+
+        // для документов, контролируем заполненность даты и номера
+        let numerator;
+        if(!this._deleted) {
+          if(this instanceof DocObj || this instanceof TaskObj || this instanceof BusinessProcessObj) {
+            if(utils.blank.date == this.date) {
+              this.date = new Date();
+            }
+            if(!this.number_doc) {
+              numerator = this.new_number_doc();
+            }
+          }
+          else {
+            if(!this.id) {
+              numerator = this.new_number_doc();
+            }
+          }
+        }
+
+        // если не указаны обязательные реквизиты...
+        const {fields, tabular_sections} = this._metadata();
+        const {msg, md, cch: {properties}, classes} = this._manager._owner.$p;
+        const flds = Object.assign({}, fields);
+        if(this._manager instanceof classes.CatManager) {
+          flds.name = this._metadata('name') || {};
+          flds.id = this._metadata('id') || {};
+        }
+        for (const mf in flds) {
+          if (flds[mf] && flds[mf].mandatory && (!this._obj[mf] || this._obj[mf] === utils.blank.guid)) {
+            return reset_mandatory({
+              obj: this,
+              title: msg.mandatory_title,
+              type: 'alert-error',
+              text: msg.mandatory_field.replace('%1', this._metadata(mf).synonym)
+            });
+          }
+        }
+        if(properties) {
+          for (const prts of ['extra_fields', 'product_params', 'params']) {
+            if(!tabular_sections[prts]) {
+              continue;
+            }
+            for (const row of this[prts]._obj) {
+              const property = properties.get(row.property || row.param);
+              if(property && property.mandatory) {
+                const {value} = (row._row || row);
+                if(utils.is_data_obj(value) ? value.empty() : !value) {
+                  return reset_mandatory({
+                    obj: this,
+                    row: row._row || row,
+                    title: msg.mandatory_title,
+                    type: 'alert-error',
+                    text: msg.mandatory_field.replace('%1', property.caption || property.name)
+                  });
+                }
+              }
+            }
+          }
+        }
+
+        // в зависимости от типа кеширования, получаем saver и сохраняем объект во внешней базе
+        return (numerator || Promise.resolve())
+          .then(() => this._manager.adapter.save_obj(this, {post, operational, attachments }))
+          // и выполняем обработку после записи
+          .then(() => this.after_save())
+          .then(reset_modified)
+          .catch((err) => {
+            reset_modified();
+            throw err;
+          });
+
+      });
+
+  }
+
+
+  /**
+   * ### Возвращает присоединенный объект или файл
+   * @method get_attachment
+   * @for DataObj
+   * @param att_id {String} - идентификатор (имя) вложения
+   */
+  get_attachment(att_id) {
+    const {_manager, ref} = this;
+    return _manager.adapter.get_attachment(_manager, ref, att_id);
+  }
+
+  /**
+   * ### Сохраняет объект или файл во вложении
+   * Вызывает {{#crossLink "DataManager/save_attachment:method"}} одноименный метод менеджера {{/crossLink}} и передаёт ссылку на себя в качестве контекста
+   *
+   * @method save_attachment
+   * @for DataObj
+   * @param att_id {String} - идентификатор (имя) вложения
+   * @param attachment {Blob|String} - вложениe
+   * @param [type] {String} - mime тип
+   * @return Promise.<DataObj>
+   * @async
+   */
+  save_attachment(att_id, attachment, type) {
+    const {_manager, ref, _attachments} = this;
+    return _manager.save_attachment(ref, att_id, attachment, type)
+      .then((att) => {
+        if(!_attachments) {
+          this._attachments = {};
+        }
+        if(!this._attachments[att_id] || !att.stub) {
+          this._attachments[att_id] = att;
+        }
+        return att;
+      });
+  }
+
+
+  /**
+   * ### Удаляет присоединенный объект или файл
+   * Вызывает одноименный метод менеджера и передаёт ссылку на себя в качестве контекста
+   *
+   * @method delete_attachment
+   * @for DataObj
+   * @param att_id {String} - идентификатор (имя) вложения
+   * @async
+   */
+  delete_attachment(att_id) {
+    const {_manager, ref, _attachments} = this;
+    return _manager.delete_attachment(ref, att_id)
+      .then((att) => {
+        if(_attachments) {
+          delete _attachments[att_id];
+        }
+        return att;
+      });
+  }
+
+  /**
+   * Применяет атрибуты к объекту
+   * @param attr
+   * @private
+   */
+  _mixin(attr, include, exclude, silent) {
+    if(Object.isFrozen(this)) {
+      return;
+    }
+    if(attr && typeof attr == 'object') {
+      const {_not_set_loaded} = attr;
+      _not_set_loaded && delete attr._not_set_loaded;
+      const {_data} = this;
+      if(silent) {
+        if(_data._loading) {
+          silent = false;
+        }
+        _data._loading = true;
+      }
+      utils._mixin(this, attr, include, exclude);
+      if(silent) {
+        _data._loading = false;
+      }
+      if(!_not_set_loaded && (_data._loading || (!utils.is_empty_guid(this.ref) && (attr.id || attr.name || attr.number_doc)))) {
+        this._set_loaded(this.ref);
+      }
+    }
+  }
+
+  /**
+   * ### Приводит строки дат к датам, ссылки к ссылкам
+   * @param obj
+   * @param _obj
+   * @param fields
+   */
+  static fix_collection(obj, _obj, fields) {
+    for (const fld in fields) {
+      if(_obj[fld]) {
+        let {type, choice_type} = fields[fld];
+        if(choice_type && choice_type.path){
+          const prop = obj[choice_type.path[choice_type.path.length - 1]];
+          if(prop && prop.type) {
+            type = prop.type;
+          }
+        }
+        if (type.is_ref && typeof _obj[fld] === 'object') {
+          if(!(fld === 'type' && obj.class_name && obj.class_name.indexOf('cch.') === 0)) {
+            _obj[fld] = utils.fix_guid(_obj[fld], false);
+          }
+        }
+        else if (type.date_part && typeof _obj[fld] === 'string') {
+          _obj[fld] = utils.fix_date(_obj[fld], type.types.length === 1);
+        }
+      }
+    }
+  }
+
+  /**
+   * ### Приводит строки дат к датам, ссылки к ссылкам в реквизитах объекта, создаёт табчасти
+   */
+  _fix_plain() {
+    const {_obj, _manager} = this;
+    const {fields, tabular_sections, hierarchical, has_owners} = this._metadata();
+
+    // корректируем реквизиты
+    DataObj.fix_collection(this, _obj, fields);
+
+    // корректируем системные реквизиты
+    if(hierarchical || has_owners) {
+      const sys_fields = {};
+      if(hierarchical) {
+        sys_fields.parent = this._metadata('parent');
+      }
+      if(has_owners) {
+        sys_fields.owner = this._metadata('owner');
+      }
+      DataObj.fix_collection(this, _obj, sys_fields);
+    }
+
+    for (const ts in tabular_sections) {
+      if(Array.isArray(_obj[ts])){
+        const tabular = this[ts];
+        const Constructor = _manager.obj_constructor(ts, true);
+        const {fields} = tabular_sections[ts];
+        for(let i = 0; i < _obj[ts].length; i++) {
+          const row = _obj[ts][i];
+          const _row = new Constructor(tabular, row);
+          row.row = i + 1;
+          Object.defineProperty(row, '_row', {value: _row});
+          DataObj.fix_collection(_row, row, fields);
+        }
+      }
+    }
+  }
+
+
+  /**
+   * ### Выполняет команду печати
+   * Вызывает одноименный метод менеджера и передаёт себя в качестве объекта печати
+   *
+   * @method print
+   * @for DataObj
+   * @param model {String} - идентификатор макета печатной формы
+   * @param [wnd] - указатель на форму, из которой произведён вызов команды печати
+   * @return {*|{value}|void}
+   * @async
+   */
+  print(model, wnd) {
+    return this._manager.print(this, model, wnd);
+  }
+
+  /**
+   * ### После создания
+   * Возникает после создания объекта. В обработчике можно установить значения по умолчанию для полей и табличных частей
+   * или заполнить объект на основании данных связанного объекта
+   *
+   * @event AFTER_CREATE
+   */
+  after_create() {
+    return this;
+  }
+
+  /**
+   * ### После чтения объекта с сервера
+   * Имеет смысл для объектов с типом кеширования ("doc", "remote", "meta", "e1cib").
+   * т.к. структура _DataObj_ может отличаться от прототипа в базе-источнике, в обработчике можно дозаполнить или пересчитать реквизиты прочитанного объекта
+   *
+   * @event AFTER_LOAD
+   */
+  after_load() {
+    return this;
+  }
+
+  /**
+   * ### Перед записью
+   * Возникает перед записью объекта. В обработчике можно проверить корректность данных, рассчитать итоги и т.д.
+   * Запись можно отклонить, если у пользователя недостаточно прав, либо введены некорректные данные
+   *
+   * @event BEFORE_SAVE
+   */
+  before_save() {
+    return this;
+  }
+
+  /**
+   * ### После записи
+   *
+   * @event AFTER_SAVE
+   */
+  after_save() {
+    return this;
+  }
+
+  /**
+   * ### При изменении реквизита шапки или табличной части
+   *
+   * @event VALUE_CHANGE
+   */
+  value_change(f, mf, v) {
+    return this;
+  }
+
+  /**
+   * ### При добавлении строки табличной части
+   *
+   * @event ADD_ROW
+   */
+  add_row(row) {
+    return this;
+  }
+
+  /**
+   * ### При удалении строки табличной части
+   *
+   * @event DEL_ROW
+   */
+  del_row(row) {
+    return this;
+  }
+
+  /**
+   * ### После удаления строки табличной части
+   *
+   * @event AFTER_DEL_ROW
+   */
+  after_del_row(name) {
+    return this;
+  }
 
 }
 
@@ -606,13 +868,19 @@ class DataObj {
  * @for DataObj
  * @type String
  */
-Object.defineProperty(DataObj.prototype, "ref", {
-	get : function(){ return this._obj.ref},
-	set : function(v){ this._obj.ref = utils.fix_guid(v)},
-	enumerable : true,
-	configurable: true
-})
+Object.defineProperty(DataObj.prototype, 'ref', {
+  get: function () {
+    return this._obj ? this._obj.ref : utils.blank.guid;
+  },
+  set: function (v) {
+    this._obj.ref = utils.fix_guid(v);
+  },
+  enumerable: true,
+  configurable: true
+});
 
+TabularSectionRow.prototype._getter = DataObj.prototype._getter;
+TabularSectionRow.prototype.__setter = DataObj.prototype.__setter;
 
 
 /**
@@ -623,71 +891,155 @@ Object.defineProperty(DataObj.prototype, "ref", {
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {RefDataManager}
  */
-class CatObj extends DataObj{
+export class CatObj extends DataObj {
 
-	constructor(attr, manager){
+  constructor(attr, manager, loading) {
 
-		// выполняем конструктор родительского объекта
-		super(attr, manager);
+    const direct = loading && attr && utils.is_guid(attr.ref);
 
-		this._mixin_attr(attr);
+    // выполняем конструктор родительского объекта
+    super(attr, manager, loading, direct);
 
-	}
+    if(direct) {
+      this._fix_plain();
+    }
+    else {
+      this._mixin(attr);
+    }
 
-	/**
-	 * Представление объекта
-	 * @property presentation
-	 * @for CatObj
-	 * @type String
-	 */
-	get presentation(){
-		if(this.name || this.id){
-			// return this._metadata.obj_presentation || this._metadata.synonym + " " + this.name || this.id;
-			return this.name || this.id || this._metadata.obj_presentation || this._metadata.synonym;
-		}else
-			return this._presentation || '';
-	}
-	/**
-	 * @type String
-	 */
-	set presentation(v){
-		if(v)
-			this._presentation = String(v);
-	}
+  }
+
+  /**
+   * Представление объекта
+   * @property presentation
+   * @for CatObj
+   * @type String
+   */
+  get presentation() {
+    return this.name || this.id || this._presentation || '';
+  }
+  set presentation(v) {
+    if(v) {
+      this._presentation = String(v);
+    }
+  }
+
+  /**
+   * ### Код элемента справочника
+   * @property id
+   * @type String|Number
+   */
+  get id() {
+    return this._obj.id || '';
+  }
+  set id(v) {
+    this.__notify('id');
+    this._obj.id = v;
+  }
+
+  /**
+   * ### Наименование элемента справочника
+   * @property name
+   * @type String
+   */
+  get name() {
+    return this._obj.name || '';
+  }
+  set name(v) {
+    this.__notify('name');
+    this._obj.name = String(v);
+  }
+
+
+  /**
+   * ### Дети
+   * Возвращает массив элементов, находящихся в иерархии текущего
+   */
+  _children(folders) {
+    const res = [];
+    this._manager.forEach((o) => {
+      if(o != this && (!folders || o.is_folder) && o._hierarchy(this)) {
+        res.push(o);
+      }
+    });
+    return res;
+  }
+
+  /**
+   * ### Родители
+   * Возвращает массив родителей, в иерархии которых находится текущий элемент
+   * @private
+   */
+  _parents() {
+    const res = [];
+    let {parent} = this;
+    while (parent && !parent.empty()) {
+      res.push(parent);
+      parent = parent.parent;
+    }
+    return res;
+  }
+
+  /**
+   * ### В иерархии
+   * Выясняет, находится ли текущий объект в указанной группе
+   *
+   * @param group {Object|Array} - папка или массив папок
+   *
+   */
+  _hierarchy(group) {
+    if(Array.isArray(group)) {
+      return group.some((v) => this._hierarchy(v));
+    }
+    const {parent} = this;
+    if(this == group || parent == group) {
+      return true;
+    }
+    if(parent && !parent.empty()) {
+      return parent._hierarchy(group);
+    }
+    return group == utils.blank.guid;
+  }
+
 
 }
-Object.defineProperties(CatObj.prototype, {
 
-	/**
-	 * ### Код элемента справочника
-	 * @property id
-	 * @type String|Number
-	 */
-	id: {
-		get : function(){ return this._obj.id || ""},
-		set : function(v){
-			this.__notify('id');
-			this._obj.id = v;
-		},
-		enumerable: true
-	},
+/**
+ * mixin свойств дата и номер документа к базовому классу
+ * @param superclass
+ * @constructor
+ */
+export const NumberDocAndDate = (superclass) => class extends superclass {
 
-	/**
-	 * ### Наименование элемента справочника
-	 * @property name
-	 * @type String
-	 */
-	name: {
-		get : function(){ return this._obj.name || ""},
-		set : function(v){
-			this.__notify('name');
-			this._obj.name = String(v);
-		},
-		enumerable: true
-	}
-})
+  /**
+   * Номер документа
+   * @property number_doc
+   * @type {String|Number}
+   */
+  get number_doc() {
+    return this._obj.number_doc || '';
+  }
 
+  set number_doc(v) {
+    this.__notify('number_doc');
+    this._obj.number_doc = v;
+  }
 
+  /**
+   * Дата документа
+   * @property date
+   * @type {Date}
+   */
+  get date() {
+    return this._obj.date instanceof Date ? this._obj.date : utils.blank.date;
+  }
+
+  set date(v) {
+    this.__notify('date');
+    this._obj.date = utils.fix_date(v, true);
+  }
+
+};
 
 /**
  * ### Абстрактный класс ДокументОбъект
@@ -697,89 +1049,60 @@ Object.defineProperties(CatObj.prototype, {
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {RefDataManager}
  */
-class DocObj extends DataObj{
+export class DocObj extends NumberDocAndDate(DataObj) {
 
-	constructor(attr, manager){
+  constructor(attr, manager, loading) {
 
-		// выполняем конструктор родительского объекта
-		super(attr, manager);
+    const direct = loading && attr && utils.is_guid(attr.ref);
 
-		this._mixin_attr(attr);
+    // выполняем конструктор родительского объекта
+    super(attr, manager, loading, direct);
 
-	}
+    if(direct) {
+      this._fix_plain(this);
+    }
+    else {
+      this._mixin(attr);
+    }
 
-	/**
-	 * Представление объекта
-	 * @property presentation
-	 * @for DocObj
-	 * @type String
-	 */
-	get presentation(){
-		if(this.number_doc)
-			return (this._metadata.obj_presentation || this._metadata.synonym) + ' №' + this.number_doc + " от " + moment(this.date).format(moment._masks.ldt);
-		else
-			return this._presentation || "";
-	}
-	/**
-	 * @type String
-	 */
-	set presentation(v){
-		if(v)
-			this._presentation = String(v);
-	}
+  }
+
+  /**
+   * Представление объекта
+   * @property presentation
+   * @for DocObj
+   * @type String
+   */
+  get presentation() {
+    const meta = this._metadata();
+    const {number_doc, date, posted, _modified} = this;
+    return number_doc ?
+      `${meta.obj_presentation || meta.synonym}  №${number_doc} от ${moment(date).format(moment._masks.date_time)} (${posted ? '' : 'не '}проведен)${_modified ? ' *' : ''}`
+      :
+      `${meta.obj_presentation || meta.synonym} ${moment(date).format(moment._masks.date_time)} (${posted ? '' : 'не '}проведен)${_modified ? ' *' : ''}`;
+  }
+
+  set presentation(v) {
+    if(v) {
+      this._presentation = String(v);
+    }
+  }
+
+  /**
+   * Признак проведения
+   * @property posted
+   * @type {Boolean}
+   */
+  get posted() {
+    return this._obj.posted || false;
+  }
+
+  set posted(v) {
+    this.__notify('posted');
+    this._obj.posted = utils.fix_boolean(v);
+  }
 
 }
-
-
-function doc_props_date_number(proto){
-
-	Object.defineProperties(proto, {
-
-		/**
-		 * Номер документа
-		 * @property number_doc
-		 * @type {String|Number}
-		 */
-		number_doc: {
-			get : function(){ return this._obj.number_doc || ""},
-			set : function(v){
-				this.__notify('number_doc');
-				this._obj.number_doc = v;
-			},
-			enumerable: true
-		},
-
-		/**
-		 * Дата документа
-		 * @property date
-		 * @type {Date}
-		 */
-		date: {
-			get : function(){ return this._obj.date || utils.blank.date},
-			set : function(v){
-				this.__notify('date');
-				this._obj.date = utils.fix_date(v, true);
-			},
-			enumerable: true
-		}
-
-	});
-}
-
-/**
- * Признак проведения
- * @property posted
- * @type {Boolean}
- */
-Object.defineProperty(DocObj.prototype, "posted", {
-	get : function(){ return this._obj.posted || false},
-	set : function(v){
-		this.__notify('posted');
-		this._obj.posted = utils.fix_boolean(v);
-	},
-	enumerable: true
-});
-doc_props_date_number(DocObj.prototype);
 
 
 /**
@@ -790,24 +1113,30 @@ doc_props_date_number(DocObj.prototype);
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {DataManager}
  */
-class DataProcessorObj extends DataObj {
+export class DataProcessorObj extends DataObj {
 
-	constructor(attr, manager) {
+  constructor(attr, manager, loading) {
 
-		// выполняем конструктор родительского объекта
-		super(attr, manager);
+    // выполняем конструктор родительского объекта
+    super(attr, manager, loading);
 
-		var f, cmd = manager.metadata();
-		for(f in cmd.fields)
-			attr[f] = utils.fetch_type("", cmd.fields[f].type);
-		for(f in cmd["tabular_sections"])
-			attr[f] = [];
+    if(!loading) {
+      const {fields, tabular_sections} = manager.metadata();
+      for (const fld in fields) {
+        if(!attr[fld]) {
+          attr[fld] = utils.fetch_type('', fields[fld].type);
+        }
+      }
+      for (const fld in tabular_sections) {
+        if(!attr[fld]) {
+          attr[fld] = [];
+        }
+      }
+    }
 
-		utils._mixin(this, attr);
-
-	}
+    utils._mixin(this, attr);
+  }
 }
-
 
 
 /**
@@ -818,8 +1147,9 @@ class DataProcessorObj extends DataObj {
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {DataManager}
  */
-class TaskObj extends CatObj {}
-doc_props_date_number(TaskObj.prototype);
+export class TaskObj extends NumberDocAndDate(CatObj) {
+
+}
 
 
 /**
@@ -830,8 +1160,9 @@ doc_props_date_number(TaskObj.prototype);
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {DataManager}
  */
-class BusinessProcessObj extends CatObj {}
-doc_props_date_number(BusinessProcessObj.prototype);
+export class BusinessProcessObj extends NumberDocAndDate(CatObj) {
+
+}
 
 
 /**
@@ -844,86 +1175,94 @@ doc_props_date_number(BusinessProcessObj.prototype);
  * @param attr {Object} - объект с реквизитами в свойствах или строка guid ссылки
  * @param manager {EnumManager}
  */
-class EnumObj extends DataObj {
+export class EnumObj extends DataObj {
 
-	constructor(attr, manager) {
+  constructor(attr, manager, loading) {
 
-		// выполняем конструктор родительского объекта
-		super(attr, manager);
+    // выполняем конструктор родительского объекта
+    super(attr, manager, loading);
 
-		if(attr && typeof attr == "object")
-			utils._mixin(this, attr);
+    if(attr && typeof attr == 'object') {
+      const {_obj} = this;
+      if(!_obj.ref && _obj.name) {
+        _obj.ref = _obj.name;
+      }
+      _obj !== attr && utils._mixin(this, attr);
+    }
 
-	}
+  }
 
-	/**
-	 * Порядок элемента перечисления
-	 * @property order
-	 * @for EnumObj
-	 * @type Number
-	 */
-	get order() {
-		return this._obj.sequence
-	}
-	/**
-	 * @type Number
-	 */
-	set order(v) {
-		this._obj.sequence = parseInt(v)
-	}
+  /**
+   * Порядок элемента перечисления
+   * @property order
+   * @for EnumObj
+   * @type Number
+   */
+  get order() {
+    return this._obj.sequence;
+  }
+
+  /**
+   * @type Number
+   */
+  set order(v) {
+    this._obj.sequence = parseInt(v);
+  }
 
 
-	/**
-	 * Наименование элемента перечисления
-	 * @property name
-	 * @for EnumObj
-	 * @type String
-	 */
-	get name() {
-		return this._obj.ref
-	}
-	/**
-	 * @type String
-	 */
-	set name(v) {
-		this._obj.ref = String(v)
-	}
+  /**
+   * Наименование элемента перечисления
+   * @property name
+   * @for EnumObj
+   * @type String
+   */
+  get name() {
+    return this._obj.ref;
+  }
 
-	/**
-	 * Синоним элемента перечисления
-	 * @property synonym
-	 * @for EnumObj
-	 * @type String
-	 */
-	get synonym() {
-		return this._obj.synonym || ""
-	}
-	/**
-	 * @type String
-	 */
-	set synonym(v) {
-		this._obj.synonym = String(v)
-	}
+  /**
+   * @type String
+   */
+  set name(v) {
+    this._obj.ref = String(v);
+  }
 
-	/**
-	 * Представление объекта
-	 * @property presentation
-	 * @for EnumObj
-	 * @type String
-	 */
-	get presentation() {
-		return this.synonym || this.name;
-	}
+  /**
+   * Синоним элемента перечисления
+   * @property synonym
+   * @for EnumObj
+   * @type String
+   */
+  get synonym() {
+    return this._obj.synonym || '';
+  }
 
-	/**
-	 * Проверяет, является ли ссылка объекта пустой
-	 * @method empty
-	 * @for EnumObj
-	 * @return {boolean} - true, если ссылка пустая
-	 */
-	empty() {
-		return !this.ref || this.ref == "_";
-	}
+  /**
+   * @type String
+   */
+  set synonym(v) {
+    this._obj.synonym = String(v);
+  }
+
+  /**
+   * Представление объекта
+   * @property presentation
+   * @for EnumObj
+   * @type String
+   */
+  get presentation() {
+    return this.synonym || this.name;
+  }
+
+  /**
+   * Проверяет, является ли ссылка объекта пустой
+   * @method empty
+   * @for EnumObj
+   * @return {boolean} - true, если ссылка пустая
+   */
+  empty() {
+    return !this.ref || this.ref == '_';
+  }
 }
 
 
@@ -937,51 +1276,65 @@ class EnumObj extends DataObj {
  * @param attr {object} - объект, по которому запись будет заполнена
  * @param manager {InfoRegManager|AccumRegManager}
  */
-class RegisterRow extends DataObj {
+export class RegisterRow extends DataObj {
 
-	constructor(attr, manager) {
+  constructor(attr, manager, loading) {
 
-		// выполняем конструктор родительского объекта
-		super(attr, manager);
+    // выполняем конструктор родительского объекта
+    super(attr, manager, loading);
 
-		if (attr && typeof attr == "object")
-			utils._mixin(this, attr);
+    if(attr && typeof attr == 'object') {
+      let tref = attr.ref;
+      if(tref) {
+        delete attr.ref;
+      }
+      utils._mixin(this, attr);
+      if(tref) {
+        attr.ref = tref;
+      }
+    }
 
-		for (var check in manager.metadata().dimensions) {
-			if (!attr.hasOwnProperty(check) && attr.ref) {
-				var keys = attr.ref.split("¶");
-				Object.keys(manager.metadata().dimensions).forEach((fld, ind) => {
-					this[fld] = keys[ind]
-				});
-				break;
-			}
-		}
+    for (var check in manager.metadata().dimensions) {
+      if(!attr.hasOwnProperty(check) && attr.ref) {
+        var keys = attr.ref.split('¶');
+        Object.keys(manager.metadata().dimensions).forEach((fld, ind) => {
+          this[fld] = keys[ind];
+        });
+        break;
+      }
+    }
 
-	}
+  }
 
-	/**
-	 * Метаданные строки регистра
-	 * @property _metadata
-	 * @for RegisterRow
-	 * @type Object
-	 */
-	get _metadata() {
-		var _meta = this._manager.metadata();
-		if (!_meta.fields)
-			_meta.fields = Object.assign({}, _meta.dimensions, _meta.resources, _meta.attributes);
-		return _meta;
-	}
+  /**
+   * Метаданные строки регистра
+   * @method _metadata
+   * @for RegisterRow
+   * @param field_name
+   * @type Object
+   */
+  _metadata(field_name) {
+    const _meta = this._manager.metadata();
+    if(!_meta.fields) {
+      _meta.fields = Object.assign({}, _meta.dimensions, _meta.resources, _meta.attributes);
+    }
+    return field_name ? _meta.fields[field_name] : _meta;
+  }
 
-	/**
-	 * Ключ записи регистра
-	 */
-	get ref() {
-		return this._manager.get_ref(this);
-	}
+  /**
+   * Ключ записи регистра
+   */
+  get ref() {
+    return this._manager.get_ref(this);
+  }
 
-	get presentation() {
-		return this._metadata.obj_presentation || this._metadata.synonym;
-	}
+  set ref(v) {
+
+  }
+
+  get presentation() {
+    return this._metadata().obj_presentation || this._metadata().synonym;
+  }
 }
 
 
