@@ -40,18 +40,6 @@ function InterfaceObjs(){
 	var iface = this;
 
 	/**
-	 * Очищает область (например, удаляет из div все дочерние элементы)
-	 * @method clear_svgs
-	 * @param area {HTMLElement|String}
-	 */
-	this.clear_svgs = function(area){
-		if(typeof area === "string")
-			area = document.getElementById(area);
-		while (area.firstChild)
-			area.removeChild(area.firstChild);
-	};
-
-	/**
 	 * Возвращает координату левого верхнего угла элемента относительно документа
 	 * @method get_offset
 	 * @param elm {HTMLElement} - элемент, координату которого, необходимо определить
@@ -75,7 +63,12 @@ function InterfaceObjs(){
 	 * @return {XML|string}
 	 */
 	this.normalize_xml = function(str){
-		if(!str) return "";
+    if(!str) {
+      return '';
+    }
+    else if(typeof str === 'object') {
+      return str.toString();
+    }
 		var entities = { '&':  '&amp;', '"': '&quot;',  "'":  '&apos;', '<': '&lt;', '>': '&gt;'};
 		return str.replace(	/[&"'<>]/g, function (s) {return entities[s];});
 	};
@@ -272,12 +265,13 @@ function InterfaceObjs(){
 	 * @param e {MouseEvent|KeyboardEvent}
 	 * @returns {Boolean}
 	 */
-	this.cancel_bubble = function(e) {
+	this.cancel_bubble = function(e, prevent) {
 		var evt = (e || event);
-		if (evt && evt.stopPropagation)
-			evt.stopPropagation();
-		if (evt && !evt.cancelBubble)
-			evt.cancelBubble = true;
+    evt && prevent && evt.preventDefault && evt.preventDefault();
+		evt && evt.stopPropagation && evt.stopPropagation();
+		if (evt && !evt.cancelBubble){
+      evt.cancelBubble = true;
+    }
 		return false
 	};
 
@@ -318,7 +312,7 @@ function InterfaceObjs(){
 		// основной сайдбар
 		iface.main = new dhtmlXSideBar({
 			parent: document.body,
-			icons_path: icons_path || "dist/imgs/",
+			icons_path: icons_path || "imgs/",
 			width: 180,
 			header: true,
 			template: "tiles",
@@ -366,7 +360,7 @@ function InterfaceObjs(){
 	 */
 	function All_meta_objs(cont, classes, frm_attr) {
 
-		this.layout = cont.attachLayout({
+		var layout = this.layout = cont.attachLayout({
 			pattern: "2U",
 			cells: [{
 				id: "a",
@@ -382,18 +376,18 @@ function InterfaceObjs(){
 		});
 
 		// дерево используемых метаданных
-		this.tree = this.layout.cells("a").attachTreeView();
-		this.tree.attachEvent("onSelect", function (name, mode) {
+		var tree = this.tree = layout.cells("a").attachTreeView();
+    tree.attachEvent("onSelect", function (name, mode) {
 			if(!mode)
 				return;
 			var mgr = $p.md.mgr_by_class_name(name);
 			if(mgr instanceof DataProcessorsManager){
 				// для отчетов и обработок используем форму отчета
-				mgr.form_rep(this.layout.cells("b"), frm_attr || {hide_header: true});
+				mgr.form_rep(layout.cells("b"), frm_attr || {hide_header: true});
 
 			}else if(mgr){
 				// для остальных объектов показываем форму списка
-				mgr.form_list(this.layout.cells("b"), frm_attr || {hide_header: true});
+				mgr.form_list(layout.cells("b"), frm_attr || {hide_header: true});
 			}
 
 		}.bind(this));
@@ -414,24 +408,24 @@ function InterfaceObjs(){
 				var key = classes[0]+"."+name,
 					meta = $p.md.get(key);
 				if(!meta.hide){
-					this.tree.addItem(key, meta.list_presentation || meta.synonym);
-					this.tree.setItemIcons(key, {file: "icon_1c_"+classes[0]});
+          tree.addItem(key, meta.list_presentation || meta.synonym);
+          tree.setItemIcons(key, {file: "icon_1c_"+classes[0]});
 				}
 			}.bind(this));
 
 		}else{
 			classes.forEach(function (id) {
-				this.tree.addItem(id, $p.msg["meta_"+id]);
-				this.tree.setItemIcons(id, {file: "icon_1c_"+id, folder_opened: "icon_1c_"+id, folder_closed: "icon_1c_"+id});
+        tree.addItem(id, $p.msg["meta_"+id]);
+        tree.setItemIcons(id, {file: "icon_1c_"+id, folder_opened: "icon_1c_"+id, folder_closed: "icon_1c_"+id});
 				$p.md.get_classes()[id].forEach(function (name) {
 					var key = id+"."+name,
 						meta = $p.md.get(key);
 					if(!meta.hide){
-						this.tree.addItem(key, meta.list_presentation || meta.synonym, id);
-						this.tree.setItemIcons(key, {file: "icon_1c_"+id});
+            tree.addItem(key, meta.list_presentation || meta.synonym, id);
+            tree.setItemIcons(key, {file: "icon_1c_"+id});
 					}
-				}.bind(this));
-			}.bind(this));
+				});
+			});
 		}
 	}
 
@@ -442,166 +436,15 @@ function InterfaceObjs(){
 	 */
 	this.All_meta_objs = All_meta_objs;
 
-	/**
-	 * ### Страница общих настроек
-	 * @param cont {dhtmlXCellObject} - контейнер для размещения страницы
-	 * @constructor
-	 */
-	function Setting2col(cont) {
-
-		// закладка основных настроек
-		cont.attachHTMLString($p.injected_data['view_settings.html']);
-		this.cont = cont.cell.querySelector(".dhx_cell_cont_tabbar");
-		this.cont.style.overflow = "auto";
-
-		// первая колонка настроек
-		this.form2 = (function (cont) {
-
-			var form = new dhtmlXForm(cont, [
-
-				{ type:"settings", labelWidth:80, position:"label-left"  },
-
-				{type: "label", labelWidth:320, label: "Адрес CouchDB", className: "label_options"},
-				{type:"input" , inputWidth: 220, name:"couch_path", label:"Путь:", validate:"NotEmpty"},
-				{type:"template", label:"",value:"",
-					note: {text: "Можно указать как относительный, так и абсолютный URL публикации CouchDB", width: 320}},
-
-				{type: "label", labelWidth:320, label: "Адрес http сервиса 1С", className: "label_options"},
-				{type:"input" , inputWidth: 220, name:"rest_path", label:"Путь", validate:"NotEmpty"},
-				{type:"template", label:"",value:"",
-					note: {text: "Можно указать как относительный, так и абсолютный URL публикации 1С OData", width: 320}},
-
-				{type: "label", labelWidth:320, label: "Значение разделителя данных", className: "label_options"},
-				{type:"input" , inputWidth: 220, name:"zone", label:"Зона:", numberFormat: ["0", "", ""], validate:"NotEmpty,ValidInteger"},
-				{type:"template", label:"",value:"", note: {text: "Для неразделенной публикации, зона = 0", width: 320}},
-
-				{type: "label", labelWidth:320, label: "Суффикс базы пользователя", className: "label_options"},
-				{type:"input" , inputWidth: 220, name:"couch_suffix", label:"Суффикс:"},
-				{type:"template", label:"",value:"",
-					note: {text: "Назначается абоненту при регистрации", width: 320}},
-
-				{type:"block", blockOffset: 0, name:"block_buttons", list:[
-					{type: "button", name: "save", value: "<i class='fa fa-floppy-o fa-lg'></i>", tooltip: "Применить настройки и перезагрузить программу"},
-					{type:"newcolumn"},
-					{type: "button", offsetLeft: 20, name: "reset", value: "<i class='fa fa-refresh fa-lg'></i>", tooltip: "Стереть справочники и перезаполнить данными сервера"}
-				]
-				}
-			]);
-
-			form.cont.style.fontSize = "100%";
-
-			// инициализация свойств
-			["zone", "couch_path", "couch_suffix", "rest_path"].forEach(function (prm) {
-				if(prm == "zone")
-					form.setItemValue(prm, $p.wsql.get_user_param(prm));
-				else
-					form.setItemValue(prm, $p.wsql.get_user_param(prm) || $p.job_prm[prm]);
-			});
-
-			form.attachEvent("onChange", function (name, value, state){
-				$p.wsql.set_user_param(name, name == "enable_save_pwd" ? state || "" : value);
-			});
-
-			form.disableItem("couch_suffix");
-
-			if(!$p.job_prm.rest_path)
-				form.disableItem("rest_path");
-
-			form.attachEvent("onButtonClick", function(name){
-
-				if(name == "save"){
-
-					// завершаем синхронизацию
-					$p.wsql.pouch.log_out();
-
-					// перезагружаем страницу
-					setTimeout(function () {
-						$p.eve.redirect = true;
-						location.reload(true);
-					}, 1000);
-
-				} else if(name == "reset"){
-
-					dhtmlx.confirm({
-						title: "Сброс данных",
-						text: "Стереть справочники и перезаполнить данными сервера?",
-						cancel: $p.msg.cancel,
-						callback: function(btn) {
-							if(btn)
-								$p.wsql.pouch.reset_local_data();
-						}
-					});
-				}
-			});
-
-			return form;
-
-		})(this.cont.querySelector("[name=form2]").firstChild);
-
-		// вторая колонка настроек
-		this.form1 = (function (cont) {
-
-			var form = new dhtmlXForm(cont, [
-				{ type:"settings", labelWidth:320, position:"label-left"  },
-
-				{type: "label", label: "Тип устройства", className: "label_options"},
-				{ type:"block", blockOffset: 0, name:"block_device_type", list:[
-					{ type:"settings", labelAlign:"left", position:"label-right"  },
-					{ type:"radio" , name:"device_type", labelWidth:120, label:'<i class="fa fa-desktop"></i> Компьютер', value:"desktop"},
-					{ type:"newcolumn"   },
-					{ type:"radio" , name:"device_type", labelWidth:150, label:'<i class="fa fa-mobile fa-lg"></i> Телефон, планшет', value:"phone"}
-				]  },
-				{type:"template", label:"",value:"", note: {text: "Класс устройства определяется автоматически, но пользователь может задать его явно", width: 320}},
-
-				{type: "label", label: "Сохранять пароль пользователя", className: "label_options"},
-				{type:"checkbox", name:"enable_save_pwd", label:"Разрешить:", labelWidth:90, checked: $p.wsql.get_user_param("enable_save_pwd", "boolean")},
-				{type:"template", label:"",value:"", note: {text: "Не рекомендуется, если к компьютеру имеют доступ посторонние лица", width: 320}},
-				{type:"template", label:"",value:"", note: {text: "", width: 320}},
-
-				{type: "label", label: "Подключаемые модули", className: "label_options"},
-				{type:"input" , position:"label-top", inputWidth: 320, name:"modifiers", label:"Модификаторы:", value: $p.wsql.get_user_param("modifiers"), rows: 3, style:"height:80px;"},
-				{type:"template", label:"",value:"", note: {text: "Список дополнительных модулей", width: 320}}
-
-			]);
-
-			form.cont.style.fontSize = "100%";
-
-			// инициализация свойств
-			form.checkItem("device_type", $p.job_prm.device_type);
-
-			// подключаем обработчик изменения значений в форме
-			form.attachEvent("onChange", function (name, value, state){
-				$p.wsql.set_user_param(name, name == "enable_save_pwd" ? state || "" : value);
-
-			});
-
-			form.disableItem("modifiers");
-			form.getInput("modifiers").onchange = function () {
-				$p.wsql.set_user_param("modifiers", this.value);
-			};
-
-			return form;
-
-		})(this.cont.querySelector("[name=form1]").firstChild);
-
-	}
-
-	/**
-	 * ### Страница общих настроек
-	 * @property Setting2col
-	 * @type {Setting2col}
-	 */
-	this.Setting2col = Setting2col;
-
-	this.do_reload = function () {
+	this.do_reload = function (text, title) {
 
 		var confirm_count = 0;
 
 		function do_reload(){
 
 			dhtmlx.confirm({
-				title: $p.msg.file_new_date_title,
-				text: $p.msg.file_new_date,
+				title: title || $p.msg.file_new_date_title,
+				text: text || $p.msg.file_new_date,
 				ok: "Перезагрузка",
 				cancel: "Продолжить",
 				callback: function(btn) {
@@ -628,7 +471,21 @@ function InterfaceObjs(){
 
 		do_reload();
 	}
+
+  /**
+   * Проверяет, нет ли других модальных форм
+   */
+  this.check_exit = function (wnd){
+    var do_exit;
+    // если есть внешнее модальное, ничего обрабатывать не надо
+    this.w.forEachWindow(function (w) {
+      if(w != wnd && (w.isModal() || $p.iface.w.getTopmostWindow() == w))
+        do_exit = true;
+    });
+    return do_exit;
+  }
 }
+
 
 $p.__define({
 
@@ -653,86 +510,48 @@ $p.__define({
 	 */
 	current_user: {
 		get: function () {
-			return $p.cat && $p.cat.users ?
-				$p.cat.users.by_id($p.wsql.get_user_param("user_name")) :
-				$p.utils.blank.guid;
-		}
-	},
 
-	/**
-	 * ### Права доступа текущего пользователя.
-	 * Свойство определено после загрузки метаданных и входа впрограмму
-	 * @property current_acl
-	 * @type CatUsers_acl
-	 * @final
-	 */
-	current_acl: {
-		get: function () {
+      if($p.CatUsers && !$p.CatUsers.prototype.hasOwnProperty("role_available")){
 
-			var res, proto;
-
-			if($p.cat && $p.cat.users_acl){
-
-				$p.cat.users_acl.find_rows({owner: $p.current_user}, function (o) {
-					res = o;
-					return false;
-				});
-
-				proto = $p.CatUsers_acl.prototype;
-			}
-
-			if(!proto){
-				proto = {};
-			}
-
-			if(!res) {
-				if(this.utils.blank.users_acl){
-					res = this.utils.blank.users_acl;
-				}else{
-					res = this.utils.blank.users_acl = Object.create(proto);
-					res.__define({
-						acl_objs: {
-							value: {
-								_obj: [],
-								each: function () {},
-								find_rows: function () {}
-							}
-						}
-					});
-				}
-			}
+        $p.CatUsers.prototype.__define({
 
 
-			if(!proto.hasOwnProperty("role_available")){
-				proto.__define({
+          /**
+           * ### Роль доступна
+           *
+           * @param name {String}
+           * @returns {Boolean}
+           */
+          role_available: {
+            value: function (name) {
+              return true;
+            }
+          },
 
-					/**
-					 * ### Роль доступна
-					 *
-					 * @param name {String}
-					 * @returns {Boolean}
-					 */
-					role_available: {
-						value: function (name) {
-							return this.acl_objs._obj.some(function (row) {
-								return row.type == name;
-							});
-						}
-					},
+          /**
+           * ### Права на объект
+           * Если не задано в модификаторе, разрешаем полный доступ
+           */
+          get_acl: {
+            value: function(class_name) {
+              return "rvuidepo";
+            }
+          },
 
-					get_acl: {
-						value: function(class_name) {
-							var acn = class_name.split(".");
-							return this._acl && this._acl[acn[0]] ? this._acl[acn[0]][acn[1]] : "e";
-						}
-					}
+        });
+      }
 
-				});
-			}
+      if(!$p.cat || !$p.cat.users){
+        return $p.utils.blank.guid;
+      }
+      var res = $p.cat.users.by_id($p.wsql.get_user_param("user_name"));
+
 
 			return res;
 		}
 	},
+
+
 
 	/**
 	 * Загружает скрипты и стили синхронно и асинхронно
@@ -773,3 +592,40 @@ $p.__define({
 	}
 
 });
+
+$p.utils.__define({
+
+  /**
+   * ### Врзвращает объект Docxtemplater из blob
+   * blob может быть получен из вложения DataObj
+   *
+   * @method docxtemplater
+   * @for Utils
+   * @param blob {Blob} - двоичные данные шаблона
+   * @return {Docxtemplater} - объект open-xml-docx
+   * @async
+   */
+  docxtemplater: {
+    value: function (blob) {
+      return (window.Docxtemplater ?
+        Promise.resolve() :
+        Promise.all([
+          $p.load_script("https://cdn.jsdelivr.net/jszip/2/jszip.min.js", "script"),
+          $p.load_script("https://cdn.jsdelivr.net/combine/gh/open-xml-templating/docxtemplater-build/build/docxtemplater-latest.min.js,gh/open-xml-templating/docxtemplater-image-module-build/build/docxtemplater-image-module-latest.min.js", "script"),
+        ]))
+        .then(function () {
+          if(!Docxtemplater.prototype.saveAs){
+            Docxtemplater.prototype.saveAs = function (name) {
+              var out = this.getZip().generate({type: "blob", mimeType: $p.utils.mime_lookup('docx')});
+              $p.wsql.alasql.utils.saveAs(out, name);
+            };
+          }
+          return $p.utils.blob_as_text(blob, 'array');
+        })
+        .then(function (buffer) {
+          return new Docxtemplater().loadZip(new JSZip(buffer));
+        });
+    }
+  }
+
+})
