@@ -20,7 +20,7 @@ function Messages(){
 
 	this.toString = function(){return "Интернационализация сообщений"};
 
-	
+
 	/**
 	 * расширяем мессанджер
 	 */
@@ -38,8 +38,9 @@ function Messages(){
 		 *      text:"Error"});
 		 */
 		this.show_msg = function(attr, delm){
-			if(!attr)
-				return;
+			if(!attr){
+        return;
+      }
 			if(typeof attr == "string"){
 				if($p.iface.synctxt){
 					$p.iface.synctxt.show_message(attr);
@@ -47,10 +48,15 @@ function Messages(){
 				}
 				attr = {type:"info", text:attr };
 			}
-			if(delm && typeof delm.setText == "function")
-				delm.setText(attr.text);
+			else if(Array.isArray(attr) && attr.length > 1){
+        attr = {type: "info", text: '<b>' + attr[0] + '</b><br />' + attr[1]};
+      }
+			if(delm && typeof delm.setText == "function"){
+        delm.setText(attr.text);
+      }
 			dhtmlx.message(attr);
 		};
+    dhtmlx.message.position = "bottom";
 
 		/**
 		 * Проверяет корректность ответа сервера
@@ -133,47 +139,6 @@ if(typeof window !== "undefined" && window.dhx4){
 	};
 }
 
-/**
- *  строки ФИАС адресного классификатора
- */
-$p.fias = function FIAS(){};
-(function (fias){
-
-	fias.toString = function(){return "Коды адресного классификатора"};
-
-	fias.types = ["владение", "здание", "помещение"];
-
-	// Код, Наименование, Тип, Порядок, КодФИАС
-	fias["1010"] = {name: "дом",			type: 1, order: 1, fid: 2, syn: [" д.", " д ", " дом"]};
-	fias["1020"] = {name: "владение",		type: 1, order: 2, fid: 1, syn: [" вл.", " вл ", " влад.", " влад ", " владен.", " владен ", " владение"]};
-	fias["1030"] = {name: "домовладение",	type: 1, order: 3, fid: 3};
-
-	fias["1050"] = {name: "корпус",		type: 2, order: 1, syn: [" к.", " к ", " корп.", " корп ", "корпус"]};
-	fias["1060"] = {name: "строение",	type: 2, order: 2, fid: 1, syn: [" стр.", " стр ", " строен.", " строен ", "строение"]};
-	fias["1080"] = {name: "литера",		type: 2, order: 3, fid: 3, syn: [" л.", " л ", " лит.", " лит ", "литера"]};
-	fias["1070"] = {name: "сооружение",	type: 2, order: 4, fid: 2, syn: [" соор.", " соор ", " сооруж.", " сооруж ", "сооружение"]};
-	fias["1040"] = {name: "участок",	type: 2, order: 5, syn: [" уч.", " уч ", "участок"]};
-
-	fias["2010"] = {name: "квартира",	type: 3, order: 1, syn: ["кв.", "кв ", "кварт.", "кварт ", "квартира", "-"]};
-	fias["2030"] = {name: "офис",		type: 3, order: 2, syn: ["оф.", "оф ", "офис", "-"]};
-	fias["2040"] = {name: "бокс",		type: 3, order: 3};
-	fias["2020"] = {name: "помещение",	type: 3, order: 4};
-	fias["2050"] = {name: "комната",	type: 3, order: 5, syn: ["комн.", "комн ", "комната"]};
-
-	//	//  сокращения 1C для поддержки обратной совместимости при парсинге
-	//	fias["2010"] = {name: "кв.",	type: 3, order: 6};
-	//	fias["2030"] = {name: "оф.",	type: 3, order: 7};
-
-	// Уточняющие объекты
-	fias["10100000"] = {name: "Почтовый индекс"};
-	fias["10200000"] = {name: "Адресная точка"};
-	fias["10300000"] = {name: "Садовое товарищество"};
-	fias["10400000"] = {name: "Элемент улично-дорожной сети, планировочной структуры дополнительного адресного элемента"};
-	fias["10500000"] = {name: "Промышленная зона"};
-	fias["10600000"] = {name: "Гаражно-строительный кооператив"};
-	fias["10700000"] = {name: "Территория"};
-
-})($p.fias);
 
 /**
  * Строки сообщений и элементов интерфейса
@@ -260,8 +225,8 @@ $p.fias = function FIAS(){};
 	msg.meta_task_mgr = "Менеджер задач";
 	msg.meta_bp_mgr = "Менеджер бизнес-процессов";
 	msg.meta_reports_mgr = "Менеджер отчетов";
-	msg.meta_charts_of_accounts_mgr = "Менеджер планов счетов";
-	msg.meta_charts_of_characteristic_mgr = "Менеджер планов видов характеристик";
+	msg.meta_cacc_mgr = "Менеджер планов счетов";
+	msg.meta_cch_mgr = "Менеджер планов видов характеристик";
 	msg.meta_extender = "Модификаторы объектов и менеджеров";
 
 	msg.modified_close = "Объект изменен<br/>Закрыть без сохранения?";
@@ -290,7 +255,7 @@ $p.fias = function FIAS(){};
 	msg.select_grp = "Укажите группу, а не элемент";
 	msg.select_elm = "Укажите элемент, а не группу";
 	msg.select_file_import = "Укажите файл для импорта";
-		
+
 	msg.srv_overload = "Сервер перегружен";
 	msg.sub_row_change_disabled = "Текущая строка подчинена продукции.<br/>Строку нельзя изменить-удалить в документе<br/>только через построитель";
 	msg.sync_script = "Обновление скриптов приложения:";
