@@ -20,25 +20,24 @@ var eXcell_proto = new eXcell();
 eXcell_proto.input_keydown = function(e, t){
 
 	function obj_on_select(v){
-		if(t.source.on_select)
-			t.source.on_select.call(t.source, v);
+    t.source && t.source.on_select && t.source.on_select.call(t.source, v);
 	}
 
 	if(e.keyCode === 8 || e.keyCode === 46){          // по {del} и {bs} очищаем значение
 		t.setValue("");
 		t.grid.editStop();
-		if(t.source.on_select)
-			t.source.on_select.call(t.source, "");
-
-	}else if(e.keyCode === 9 || e.keyCode === 13)
+    t.source && t.source.on_select && t.source.on_select.call(t.source, "");
+	}
+	else if(e.keyCode === 9 || e.keyCode === 13)
 		t.grid.editStop();                          // по {tab} и {enter} заканчиваем редактирование
 
 	else if(e.keyCode === 115)
-		t.cell.firstChild.childNodes[1].onclick(e); // по {F4} открываем редактор
+		t.cell.firstChild.childNodes[1].onclick(e); // по {F4} открываем форму списка
 
 	else if(e.keyCode === 113){                      // по {F2} открываем форму объекта
 		if(t.source.tabular_section){
-			t.mgr = _md.value_mgr(t.source.row, t.source.col, t.source.row._metadata.fields[t.source.col].type);
+			t.mgr = _md.value_mgr(t.source.row, t.source.col, typeof t.source.row._metadata == 'function' ?
+        t.source.row._metadata(t.source.col).type : t.source.row._metadata.fields[t.source.col].type);
 			if(t.mgr){
 				var tv = t.source.row[t.source.col];
 				t.mgr.form_obj(t.source.wnd, {
@@ -48,7 +47,8 @@ eXcell_proto.input_keydown = function(e, t){
 			}
 
 		}else if(t.fpath.length==1){
-			t.mgr = _md.value_mgr(t.source.o._obj, t.fpath[0], t.source.o._metadata.fields[t.fpath[0]].type);
+			t.mgr = _md.value_mgr(t.source.o._obj, t.fpath[0], typeof t.source.o._metadata == 'function' ?
+        t.source.o._metadata(t.fpath[0]).type : t.source.o._metadata.fields[t.fpath[0]].type);
 			if(t.mgr){
 				var tv = t.source.o[t.fpath[0]];
 				t.mgr.form_obj(t.source.wnd, {
@@ -109,10 +109,16 @@ function eXcell_ocombo(cell){
 		t.val = t.getValue();		//save current value
 		t.cell.innerHTML = "";
 		t.combo = new OCombo({
-			parent: t.cell
+			parent: t.cell,
+			grid: t.grid
 		}._mixin(t.grid.get_cell_field()));
 		t.combo.getInput().focus();
 	};
+
+  t.open_selection = function () {
+    t.edit();
+    t.combo && t.combo.open_selection && t.combo.open_selection();
+  }
 
 	/**
 	 * вызывается при отключении редактора
@@ -241,12 +247,12 @@ eXcell_dhxCalendar.prototype.setCValue = function(val, val2){
 			if($p.ajax.authorized){
 				username = $p.ajax.username;
 				password = $p.aes.Ctr.decrypt($p.ajax.password);
-				
+
 			}else{
 				if($p.job_prm.guest_name){
 					username = $p.job_prm.guest_name;
 					password = $p.aes.Ctr.decrypt($p.job_prm.guest_pwd);
-					
+
 				}else{
 					username = $p.wsql.get_user_param("user_name");
 					password = $p.aes.Ctr.decrypt($p.wsql.get_user_param("user_pwd"));
@@ -409,7 +415,8 @@ $p.iface.data_to_grid = function (data, attr){
 	}
 
 	var xml = "<?xml version='1.0' encoding='UTF-8'?><rows total_count='%1' pos='%2' set_parent='%3'>"
-			.replace("%1", attr._total_count || data.length).replace("%2", attr.start)
+			.replace("%1", attr._total_count || data.length)
+      .replace("%2", attr.start)
 			.replace("%3", attr.set_parent || "" ),
 		caption = this.caption_flds(attr);
 
@@ -439,13 +446,13 @@ $p.iface.data_to_tree = function (data) {
 	function add_hierarchically(arr, row){
 		var curr = {id: row.ref, text: row.presentation, items: []};
 		arr.push(curr);
-		$p._find_rows(data, {parent: row.ref}, function(r){
+		$p.utils._find_rows(data, {parent: row.ref}, function(r){
 			add_hierarchically(curr.items, r);
 		});
 		if(!curr.items.length)
 			delete curr.items;
 	}
-	$p._find_rows(data, {parent: $p.utils.blank.guid}, function(r){
+	$p.utils._find_rows(data, {parent: $p.utils.blank.guid}, function(r){
 		add_hierarchically(res, r);
 	});
 
