@@ -32,17 +32,22 @@
  */
 dhtmlXCellObject.prototype.attachTabular = function(attr) {
 
-
 	var _obj = attr.obj,
 		_tsname = attr.ts,
 		_ts = _obj[_tsname],
 		_mgr = _obj._manager,
 		_meta = attr.metadata || _mgr.metadata().tabular_sections[_tsname].fields,
 		_cell = this,
-		_source = {},
-		_selection = attr.selection;
-	if(!_md.ts_captions(_mgr.class_name, _tsname, _source))
-		return;
+		_source = attr.ts_captions || {},
+    _input_filter = "",
+    _input_filter_changed = 0,
+		_selection = attr.selection || {};
+
+	if(!attr.ts_captions && !_md.ts_captions(_mgr.class_name, _tsname, _source)){
+    return;
+  }
+
+  _selection.filter_selection = filter_selection;
 
 	var _grid = this.attachGrid(),
 		_toolbar = this.attachToolbar(),
@@ -62,9 +67,22 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 	 * добавляет строку табчасти
 	 */
 	_grid._add_row = function(){
-		if(!attr.read_only){
+		if(!attr.read_only && !attr.disable_add_del){
 
-			var row = _ts.add();
+			var proto;
+			if(_selection){
+				for(var sel in _selection){
+					if(!_meta[sel] || (typeof _selection[sel] == 'function') || (typeof _selection[sel] == 'object' && !$p.is_data_obj(_selection[sel]))){
+						continue;
+					}
+					if(!proto){
+						proto = {};
+					}
+					proto[sel] = _selection[sel];
+				}
+			}
+
+			var row = _ts.add(proto);
 
 			if(_mgr.handle_event(_obj, "add_row",
 					{
@@ -81,16 +99,40 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 		}
 	};
 
+  _grid._move_row = function(direction){
+    if(attr.read_only){
+      return;
+    }
+    const r0 = get_sel_index();
+
+    if(r0 != undefined){
+      const r1 = get_sel_index(true, direction);
+
+      if(direction == "up"){
+        if(r0 >= 0 && r1 >= 0){
+          _ts.swap(r1, r0);
+          setTimeout(() => _grid.selectRow(r1, true), 100)
+        }
+      }
+      else{
+        if(r0 < _ts.count() && r1 < _ts.count()){
+          _ts.swap(r0, r1);
+          setTimeout(() => _grid.selectRow(r1, true), 100)
+        }
+      }
+    }
+  }
+
 	/**
 	 * удаляет строку табчасти
 	 */
-	_grid._del_row = function(){
-		if(!attr.read_only){
-			var rId = get_sel_index();
+	_grid._del_row = function(keydown){
+
+		if(!attr.read_only && !attr.disable_add_del){
+			const rId = get_sel_index();
 
 			if(rId != undefined){
-
-				if(_mgr.handle_event(_obj, "del_row", 
+				if(_mgr.handle_event(_obj, "del_row",
 						{
 							tabular_section: _tsname,
 							grid: _grid,
@@ -102,28 +144,44 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 				_ts.del(rId);
 
 				setTimeout(function () {
-					_grid.selectRowById(rId < _ts.count() ? rId + 1 : rId);
+					_grid.selectRowById(rId < _ts.count() ? rId + 1 : rId, true);
 				}, 100);
 			}
 		}
 	};
-	
 
-	function get_sel_index(silent){
-		var selId = _grid.getSelectedRowId();
 
-		if(selId && !isNaN(Number(selId)))
-			return Number(selId)-1;
+	function get_sel_index(silent, direction){
+    let selId = parseFloat(_grid.getSelectedRowId());
 
-		if(!silent)
-			$p.msg.show_msg({
-				type: "alert-warning",
-				text: $p.msg.no_selected_row.replace("%1", _obj._metadata.tabular_sections[_tsname].synonym || _tsname),
-				title: (_obj._metadata.obj_presentation || _obj._metadata.synonym) + ': ' + _obj.presentation
-			});
+		if(!isNaN(selId)){
+
+      if(direction == "up"){
+        selId -= 1;
+        while (!_grid.getRowById(selId) && selId >= 0) {
+          selId -= 1;
+        }
+      }
+      else if(direction){
+        selId += 1;
+        while (!_grid.getRowById(selId) && selId < _ts.count()) {
+          selId += 1;
+        }
+      }
+
+      return direction ? ((_grid.getRowById(selId) || undefined) && selId - 1) : (selId - 1);
+    }
+
+		if(!silent){
+      const _tssynonym = (typeof _obj._metadata == 'function' ? _obj._metadata(_tsname) : _obj._metadata.tabular_sections[_tsname]).synonym;
+      $p.msg.show_msg({
+        type: "alert-warning",
+        text: $p.msg.no_selected_row.replace("%1", _tssynonym || _tsname),
+        title: (_meta.obj_presentation || _meta.synonym) + ': ' + _obj.presentation
+      });
+    }
 	}
 
-	
 
 	/**
 	 * обработчик изменения значения примитивного типа
@@ -133,8 +191,11 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 		if(stage != 2 || nValue == oValue)
 			return true;
 
-		var cell_field = _grid.get_cell_field(),
-			ret_code = _mgr.handle_event(_obj, "value_change", {
+		var cell_field = _grid.get_cell_field();
+		if(!cell_field){
+      return true;
+    }
+		var	ret_code = _mgr.handle_event(_obj, "value_change", {
 				field: cell_field.field,
 				value: nValue,
 				tabular_section: _tsname,
@@ -152,17 +213,12 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 	}
 
 	/**
-	 * наблюдатель за изменениями насбор строк табчасти
+	 * наблюдатель за изменениями строк табчасти
 	 * @param changes
 	 */
-	function observer_rows(changes){
-		if(_grid.clearAll){
-			changes.some(function(change){
-				if (change.type == "rows" && change.tabular == _tsname){
-					_ts.sync_grid(_grid, _selection);
-					return true;
-				}
-			});	
+	function listener_rows(obj, fields){
+		if(_obj === obj && fields[_tsname] && _grid.clearAll){
+      _ts.sync_grid(_grid, _selection);
 		}
 	}
 
@@ -170,21 +226,24 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 	 * наблюдатель за изменениями значений в строках табчасти
 	 * @param changes
 	 */
-	function observer(changes){
-		if(changes.length > 20){
-			try{_ts.sync_grid(_grid, _selection);} catch(err){}
-		} else
-			changes.forEach(function(change){
-				if (_tsname == change.tabular){
-					if(!change.row || _grid.getSelectedRowId() != change.row.row)
-						_ts.sync_grid(_grid, _selection);
-					else{
-						if(_grid.getColIndexById(change.name) != undefined)
-							_grid.cells(change.row.row, _grid.getColIndexById(change.name))
-								.setCValue($p.utils.is_data_obj(change.row[change.name]) ? change.row[change.name].presentation : change.row[change.name]);
-					}
-				}
-			});
+	function listener(obj, fields){
+	  if(obj._owner !== _ts){
+	    return;
+    }
+    const {row} = obj;
+    if(_grid.getSelectedRowId() != row)
+      _ts.sync_grid(_grid, _selection);
+    else{
+      const cc = _grid.getColumnCount();
+      for(let i = 0; i < cc; i++){
+        if(!_grid.isColumnHidden(i)){
+          const val = obj[_grid.getColumnId(i)];
+          if(typeof val != "boolean"){
+            _grid.cells(row, i).setCValue(val.toString());
+          }
+        }
+      }
+    }
 	}
 
 	function onpaste(e) {
@@ -205,11 +264,83 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 		}
 	}
 
+  function filter_selection(row) {
+
+	  if(!_input_filter){
+	    return true;
+    }
+
+    var res;
+    _source.fields.some(function (fld) {
+      var v = row._row[fld];
+      if($p.utils.is_data_obj(v)){
+        if(!v.is_new() && v.presentation.match(_input_filter)){
+          return res = true;
+        }
+      }
+      else if(typeof v == 'number'){
+        return res = v.toLocaleString().match(_input_filter);
+      }
+      else if(v instanceof Date){
+        return res = $p.moment(v).format($p.moment._masks.date_time).match(_input_filter);
+      }
+      else if(v.match){
+        return res = v.match(_input_filter);
+      }
+    })
+    return res;
+  }
+
+	function filter_change() {
+
+    if(_input_filter_changed){
+      clearTimeout(_input_filter_changed);
+      _input_filter_changed = 0;
+    }
+
+    if(_input_filter != _cell.input_filter.value){
+      _input_filter = new RegExp(_cell.input_filter.value, 'i');
+      listener_rows(_obj, {[_tsname]: true});
+    }
+
+  }
+
+  function filter_click() {
+    var val = _cell.input_filter.value;
+    setTimeout(function () {
+      if(val != _cell.input_filter.value)
+        filter_change();
+    })
+  }
+
+  function filter_keydown() {
+
+    if(_input_filter_changed){
+      clearTimeout(_input_filter_changed);
+    }
+
+    _input_filter_changed = setTimeout(function () {
+      if(_input_filter_changed)
+        filter_change();
+    }, 500);
+
+  }
 
 	// панель инструментов табличной части
 	_toolbar.setIconsPath(dhtmlx.image_path + 'dhxtoolbar' + dhtmlx.skin_suffix());
 	_toolbar.loadStruct(attr.toolbar_struct || $p.injected_data["toolbar_add_del.xml"], function(){
-		
+
+    this.forEachItem(function(id) {
+      if(id == "input_filter"){
+        _cell.input_filter = _toolbar.getInput(id);
+        _cell.input_filter.onchange = filter_change;
+        _cell.input_filter.onclick = filter_click;
+        _cell.input_filter.onkeydown = filter_keydown;
+        _cell.input_filter.type = "search";
+        _cell.input_filter.setAttribute("placeholder", "Фильтр");
+      }
+    })
+
 		this.attachEvent("onclick", function(btn_id){
 
 			switch(btn_id) {
@@ -221,10 +352,32 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 				case "btn_delete":
 					_grid._del_row();
 					break;
+
+        case "btn_up":
+          _grid._move_row("up");
+          break;
+
+        case "btn_down":
+          _grid._move_row("down");
+          break;
+
+      case "btn_csv":
+        _ts.export("csv");
+        break;
+
+      case "btn_json":
+        _ts.export("json");
+        break;
+
+      case "btn_xls":
+        _ts.export("xls");
+        break;
 			}
 
 		});
 	});
+
+	// поле фильтра в панели инструментов
 
 	// собственно табличная часть
 	_grid.setIconsPath(dhtmlx.image_path);
@@ -241,46 +394,70 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 	_grid.setColumnIds(_source.fields.join(","));
 	_grid.enableAutoWidth(true, 1200, 600);
 	_grid.enableEditTabOnly(true);
+	if(attr.footer){
+	  for(var fn in attr.footer){
+      fn !== 'columns' && (_grid[fn] = attr.footer[fn]);
+    }
+    _grid.attachFooter(attr.footer.columns);
+  }
 	_grid.init();
 
-	if(attr.read_only){
-		_grid.setEditable(false);
+	// гасим кнопки, если ro
+	if(attr.read_only || attr.disable_add_del){
+	  if(attr.read_only){
+      _grid.setEditable(false);
+    }
 		_toolbar.forEachItem(function (name) {
 			if(["btn_add", "btn_delete"].indexOf(name) != -1)
 				_toolbar.disableItem(name);
 		});
 	}
 
+  // добавляем кнопки сортировки, если reorder
+	if(attr.reorder){
+    var pos = _toolbar.getPosition("btn_delete");
+    if(pos){
+      _toolbar.addSeparator("sep_up", pos+1);
+      _toolbar.addButton("btn_up", pos+2, '<i class="fa fa-arrow-up fa-fw"></i>');
+      _toolbar.addButton("btn_down", pos+3, '<i class="fa fa-arrow-down fa-fw"></i>');
+      _toolbar.setItemToolTip("btn_up", "Переместить строку вверх");
+      _toolbar.setItemToolTip("btn_down", "Переместить строку вниз");
+    }
+  }
+
 	_grid.__define({
 
-		// TODO: реализовать свойство selection и его инициализацию через attr
+    _obj: {
+      get: function () {
+        return _obj;
+      }
+    },
+
 		selection: {
 			get: function () {
 				return _selection;
 			},
 			set: function (sel) {
 				_selection = sel;
-				observer_rows([{tabular: _tsname, type: "rows"}]);
+        listener_rows(_obj, {[_tsname]: true});
 			}
 		},
 
 		destructor: {
 			value: function () {
 
-				if(_obj){
-					Object.unobserve(_obj, observer);
-					Object.unobserve(_obj, observer_rows);
-				}
+			  if(_mgr){
+          _mgr.off('update', listener);
+          _mgr.off('rows', listener_rows);
+        }
 
 				_obj = null;
 				_ts = null;
 				_meta = null;
 				_mgr = null;
 				_pwnd = null;
-				_cell.detachToolbar();
-
-				_grid.entBox.removeEventListener("paste", onpaste);
-
+        _cell.detachToolbar && _cell.detachToolbar();
+        _grid.entBox && _grid.entBox.removeEventListener("paste", onpaste);
 				_destructor.call(_grid);
 			}
 		},
@@ -308,7 +485,7 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 					}
 
 					if(row && col){
-						return {obj: row, field: col}._mixin(_pwnd);
+						return {obj: row, field: col, metadata: _meta[col]}._mixin(_pwnd);
 					}
 
 				}
@@ -328,6 +505,11 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 
 	_grid.attachEvent("onEditCell", tabular_on_edit);
 
+  _grid.attachEvent("onCheck", function(rId,cInd,state){
+    _grid.selectCell(rId-1, cInd);
+    tabular_on_edit(2, rId, cInd, state);
+  });
+
 	_grid.attachEvent("onRowSelect", function(rid,ind){
 		if(_ts){
 			_grid._last = {
@@ -337,12 +519,79 @@ dhtmlXCellObject.prototype.attachTabular = function(attr) {
 		}
 	});
 
+  _grid.attachEvent("onHeaderClick", function(ind,obj){
+
+    var field = _source.fields[ind];
+    if(_grid.disable_sorting || field == 'row'){
+      return;
+    }
+    if(!_grid.sort_fields){
+      _grid.sort_fields = [];
+      _grid.sort_directions = [];
+    }
+
+    // есть ли уже такая колонка
+    var index = _grid.sort_fields.indexOf(field);
+
+    function add_field() {
+      if(index == -1){
+        _grid.sort_fields.push(field);
+        _grid.sort_directions.push("asc");
+      }
+      else{
+        if(_grid.sort_directions[index] == "asc"){
+          _grid.sort_directions[index] = "desc";
+        }
+        else{
+          _grid.sort_directions[index] = "asc";
+        }
+      }
+    }
+
+    // если кликнули с шифтом - добавляем
+    if(window.event && window.event.shiftKey){
+      add_field();
+    }
+    else{
+      if(index == -1){
+        _grid.sort_fields.length = 0;
+        _grid.sort_directions.length = 0;
+      }
+      add_field();
+    }
+
+    _ts.sort(_grid.sort_fields.map(function (field, index) {
+      return field + " " + _grid.sort_directions[index];
+    }));
+
+    _ts.sync_grid(_grid);
+
+    for(var col = 0; col < _source.fields.length; col++){
+      var field = _source.fields[col];
+      var index = _grid.sort_fields.indexOf(field);
+      if(index == -1){
+        _grid.setSortImgState(false, col);
+      }
+      else{
+        _grid.setSortImgState(true, col, _grid.sort_directions[index]);
+        setTimeout(function () {
+          if(_grid && _grid.sortImg){
+            _grid.sortImg.style.display="inline";
+          }
+        }, 200);
+        break;
+      }
+    }
+  });
+
 	// заполняем табчасть данными
-	observer_rows([{tabular: _tsname, type: "rows"}]);
+  listener_rows(_obj, {[_tsname]: true});
 
 	// начинаем следить за объектом и, его табчастью допреквизитов
-	Object.observe(_obj, observer, ["row"]);
-	Object.observe(_obj, observer_rows, ["rows"]);
+  _mgr.on({
+    update: listener,
+    rows: listener_rows,
+  });
 
 	// начинаем следить за буфером обмена
 	_grid.entBox.addEventListener('paste', onpaste);
