@@ -26,7 +26,7 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 
 	var t = this,
 		input_filter_changed = 0,
-		input_filter_width = $p.job_prm.device_type == "desktop" ? 300 : 120,
+		input_filter_width = $p.job_prm.device_type == "desktop" ? 200 : 120,
 		custom_selection = {};
 
 	if(!attr.pos)
@@ -37,17 +37,13 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 		custom_selection: {
 			get: function () {
 				return custom_selection;
-			},
-			enumerable: false,
-			configurable: false
+			}
 		},
 
 		toolbar: {
 			get: function () {
 				return attr.toolbar;
-			},
-			enumerable: false,
-			configurable: false
+			}
 		},
 
 		call_event: {
@@ -64,18 +60,21 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 
 	});
 
-	function onkeydown(){
+  function onkeydown() {
 
-		if(input_filter_changed)
-			clearTimeout(input_filter_changed);
+    if(input_filter_changed) {
+      clearTimeout(input_filter_changed);
+    }
 
-		input_filter_changed = setTimeout(function () {
-			if(input_filter_changed)
-				t.call_event();
-		}, 500);
-	}
+    if(!t.disable_timer) {
+      input_filter_changed = setTimeout(function () {
+        input_filter_changed && t._prev_input_filter != t.input_filter.value && t.call_event();
+      }, 750);
+    }
 
-	// заготовка для адаптивного фильтра
+  }
+
+  // заготовка для адаптивного фильтра
 	t.toolbar.addText("div_filter", attr.pos, "");
 	t.div = t.toolbar.objPull[t.toolbar.idPrefix + "div_filter"];
 	attr.pos++;
@@ -93,13 +92,13 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 				t.сalendar.setSensitiveRange(null, inp.value);
 		}
 
-		input_filter_width = $p.job_prm.device_type == "desktop" ? 180 : 120;
+		input_filter_width = $p.job_prm.device_type == "desktop" ? 160 : 120;
 
-		t.toolbar.addInput("input_date_from", attr.pos, "", $p.job_prm.device_type == "desktop" ? 80 : 72);
+		t.toolbar.addInput("input_date_from", attr.pos, "", $p.job_prm.device_type == "desktop" ? 78 : 72);
 		attr.pos++;
 		t.toolbar.addText("lbl_date_till", attr.pos, "-");
 		attr.pos++;
-		t.toolbar.addInput("input_date_till", attr.pos, "", $p.job_prm.device_type == "desktop" ? 80 : 72);
+		t.toolbar.addInput("input_date_till", attr.pos, "", $p.job_prm.device_type == "desktop" ? 78 : 72);
 		attr.pos++;
 
 		t.input_date_from = t.toolbar.getInput("input_date_from");
@@ -153,7 +152,9 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 
 		t.toolbar.addInput("input_filter", attr.pos, "", input_filter_width);
 		t.input_filter = t.toolbar.getInput("input_filter");
-		t.input_filter.onchange = t.call_event;
+		t.input_filter.onchange = function () {
+		  t._prev_input_filter != t.input_filter.value && t.call_event();
+    };
 		t.input_filter.onclick = function () {
 			var val = t.input_filter.value;
 			setTimeout(function () {
@@ -164,25 +165,34 @@ $p.iface.Toolbar_filter = function Toolbar_filter(attr) {
 		t.input_filter.onkeydown = onkeydown;
 		t.input_filter.type = "search";
 		t.input_filter.setAttribute("placeholder", "Фильтр");
+		if(attr.filter) {
+      t.input_filter.value = attr.filter;
+    }
 
 		t.toolbar.addSpacer("input_filter");
 
-	}else if(t.input_date_till)
-		t.toolbar.addSpacer("input_date_till");
-
-	else
-		t.toolbar.addSpacer("div_filter");
-
+	}
+  else if(t.input_date_till) {
+    t.toolbar.addSpacer("input_date_till");
+  }
+  else {
+    t.toolbar.addSpacer('div_filter');
+  }
 
 };
+
 $p.iface.Toolbar_filter.prototype.__define({
 
 	get_filter: {
 		value: function (exclude_custom) {
 
+		  if(this.input_filter){
+        this._prev_input_filter = this.input_filter.value;
+      }
+
 			var res = {
 				date_from: this.input_date_from ? $p.utils.date_add_day(dhx4.str2date(this.input_date_from.value), 0, true) : "",
-				date_till: this.input_date_till ? $p.utils.date_add_day(dhx4.str2date(this.input_date_till.value), 1, true) : "",
+				date_till: this.input_date_till ? $p.utils.date_add_day(dhx4.str2date(this.input_date_till.value), 0, true) : "",
 				filter: this.input_filter ? this.input_filter.value : ""
 			}, fld, flt;
 
@@ -202,17 +212,16 @@ $p.iface.Toolbar_filter.prototype.__define({
 
 	add_filter: {
 		value: function (elm) {
-
 			var pos = this.toolbar.getPosition("input_filter") - 2,
 				id = dhx4.newId(),
 				width = (this.toolbar.getWidth("input_filter") / 2).round(0);
-
 			this.toolbar.setWidth("input_filter", width);
 			this.toolbar.addText("lbl_"+id, pos, elm.text || "");
 			pos++;
 			this.toolbar.addInput("input_"+id, pos, "", width);
-
 			this.custom_selection[elm.name] = this.toolbar.getInput("input_"+id);
+			return this;
 		}
 	}
+
 });
