@@ -36,19 +36,19 @@ $p.iface.dat_blank = function(_dxw, attr) {
 		x: (_dxw || $p.iface.w).vp.clientWidth,
 		y: (_dxw || $p.iface.w).vp.clientHeight
 	}, _move;
-	
+
 	if(wnd_dat.getPosition()[0] + wnd_dat.getDimension()[0] > _dxw_area.x){
 		_dxw_area.x = _dxw_area.x - wnd_dat.getDimension()[0];
 		_move = true;
 	}else
 		_dxw_area.x = wnd_dat.getPosition()[0];
-	
+
 	if(wnd_dat.getPosition()[1] + wnd_dat.getDimension()[1] > _dxw_area.y){
 		_dxw_area.y = _dxw_area.y - wnd_dat.getDimension()[1];
 		_move = true;
 	}else
 		_dxw_area.y = wnd_dat.getPosition()[1];
-		
+
 	if(_move){
 		if(_dxw_area.x<0 || _dxw_area.y<0)
 			wnd_dat.maximize();
@@ -68,9 +68,9 @@ $p.iface.dat_blank = function(_dxw, attr) {
 
 	// обработчик при закрытии - анализируем модальность
 	wnd_dat.attachEvent("onClose", function () {
-		
+
 		var allow_close = typeof attr.on_close == "function" ? attr.on_close(wnd_dat) : true;
-		
+
 		if(allow_close){
 
 			// восстанавливаем модальность родительского окна
@@ -79,7 +79,7 @@ $p.iface.dat_blank = function(_dxw, attr) {
 
 			return allow_close;
 		}
-						
+
 	});
 
 	wnd_dat.setIconCss('without_icon');
@@ -132,12 +132,14 @@ $p.iface.dat_blank = function(_dxw, attr) {
 		if(attr.pwnd && attr.pwnd.setModal){
 			attr.pwnd_modal = attr.pwnd.isModal();
 			attr.pwnd.setModal(0);
-		}			
+		}
 		wnd_dat.setModal(1);
 	}
 
 	return wnd_dat;
 };
+
+
 
 /**
  * обработчик выбора значения в свойствах (ссылочные типы)
@@ -165,7 +167,7 @@ $p.iface.pgrid_on_select = function(selv){
 	}
 
 	pgrid.cells().setValue($p.utils.is_data_obj(selv) ? selv.presentation : selv || "");
-	
+
 
 	if(source.grid_on_change)
 		source.grid_on_change.call(pgrid, f, selv);
@@ -198,437 +200,6 @@ $p.iface.pgrid_on_checkbox = function(rId, cInd, state){
 
 	if(source.grid_on_change)
 		source.grid_on_change(rId, state);
-};
-
-
-function _clear_all(){
-	$p.iface.docs.__define({
-		clear_all: {
-			value: function () {
-				this.detachToolbar();
-				this.detachStatusBar();
-				this.detachObject(true);
-			},
-			enumerable: false
-		},
-		"Очистить": {
-			get: function () {
-				return this.clear_all;
-			},
-			enumerable: false
-		},
-		"Контейнер": {
-			get: function () {
-				return this.cell.querySelector(".dhx_cell_cont_layout");
-			},
-			enumerable: false
-		}
-	});
-}
-
-/**
- * Создаёт форму авторизации с обработчиками перехода к фидбэку и настройкам,
- * полем входа под гостевой ролью, полями логина и пароля и кнопкой входа
- * @method frm_auth
- * @for InterfaceObjs
- * @param attr {Object} - параметры формы
- * @param [attr.cell] {dhtmlXCellObject}
- * @return {Promise}
- */
-$p.iface.frm_auth = function (attr, resolve, reject) {
-
-	if(!attr)
-		attr = {};
-
-	var _cell, _frm, w, were_errors;
-
-	if(attr.modal_dialog){
-		if(!attr.options)
-			attr.options = {
-				name: "frm_auth",
-				caption: "Вход на сервер",
-				width: 360,
-				height: 300,
-				center: true,
-				allow_close: true,
-				allow_minmax: true,
-				modal: true
-			};
-		_cell = $p.iface.dat_blank(attr._dxw, attr.options);
-		_cell.attachEvent("onClose",function(win){
-			if(were_errors){
-				if(reject)
-					reject(err);
-			}else if(resolve)
-				resolve();
-			return true;
-		});
-		_frm = _cell.attachForm();
-
-	}else{
-		_cell = attr.cell || $p.iface.docs;
-		_frm = $p.iface.auth = _cell.attachForm();
-		$p.msg.show_msg($p.msg.init_login, _cell);
-	}
-
-
-	function do_auth(login, password){
-		
-		$p.ajax.username = login;
-		$p.ajax.password = $p.aes.Ctr.encrypt(password);
-
-		if(login){
-
-			if($p.wsql.get_user_param("user_name") != login)
-				$p.wsql.set_user_param("user_name", login);					// сохраняем имя пользователя в базе
-
-			//$p.eve.log_in(attr.onstep)
-			$p.wsql.pouch.log_in(login, password)
-				.then(function () {
-
-					if($p.wsql.get_user_param("enable_save_pwd")){
-						if($p.aes.Ctr.decrypt($p.wsql.get_user_param("user_pwd")) != password)
-							$p.wsql.set_user_param("user_pwd", $p.aes.Ctr.encrypt(password));   // сохраняем имя пользователя в базе
-						
-					}else if($p.wsql.get_user_param("user_pwd") != "")
-						$p.wsql.set_user_param("user_pwd", "");
-
-					$p.eve.logged_in = true;
-					if(attr.modal_dialog)
-						_cell.close();
-					else if(resolve)
-						resolve();
-
-				})
-				.catch(function (err) {
-					were_errors = true;
-					_frm.onerror(err);
-				})
-				.then(function () {
-					if($p.iface.sync)
-						$p.iface.sync.close();
-					if(_cell && _cell.progressOff){
-						_cell.progressOff();
-						if(!were_errors && attr.hide_header)
-							_cell.hideHeader();
-					}
-					if($p.iface.cell_tree && !were_errors)
-						$p.iface.cell_tree.expand();
-				});
-
-		} else
-			this.validate();
-	}
-
-	// обработчик кнопки "войти" формы авторизации
-	function auth_click(name){
-
-		were_errors = false;
-		this.resetValidateCss();
-
-		if(this.getCheckedValue("type") == "guest"){
-
-			var login = this.getItemValue("guest"),
-				password = "";
-			if($p.job_prm.guests && $p.job_prm.guests.length){
-				$p.job_prm.guests.some(function (g) {
-					if(g.username == login){
-						password = $p.aes.Ctr.decrypt(g.password);
-						return true;
-					}
-				});
-			}
-			do_auth.call(this, login, password);
-
-		}else if(this.getCheckedValue("type") == "auth"){
-			do_auth.call(this, this.getItemValue("login"), this.getItemValue("password"));
-
-		}
-	}
-
-	// загружаем структуру формы
-	_frm.loadStruct($p.injected_data["form_auth.xml"], function(){
-
-		var selected;
-
-		// если указан список гостевых пользователей
-		if($p.job_prm.guests && $p.job_prm.guests.length){
-
-			var guests = $p.job_prm.guests.map(function (g) {
-					var v = {
-						text: g.username,
-						value: g.username
-					};
-					if($p.wsql.get_user_param("user_name") == g.username){
-						v.selected = true;
-						selected = g.username;
-					}
-					return v;
-				});
-
-			if(!selected){
-				guests[0].selected = true;
-				selected = guests[0].value;
-			}
-
-			_frm.reloadOptions("guest", guests);
-		}
-
-		// после готовности формы читаем пользователя из локальной датабазы
-		if($p.wsql.get_user_param("user_name") && $p.wsql.get_user_param("user_name") != selected){
-			_frm.setItemValue("login", $p.wsql.get_user_param("user_name"));
-			_frm.setItemValue("type", "auth");
-
-			if($p.wsql.get_user_param("enable_save_pwd") && $p.wsql.get_user_param("user_pwd")){
-				_frm.setItemValue("password", $p.aes.Ctr.decrypt($p.wsql.get_user_param("user_pwd")));
-			}
-		}
-
-		// позиционируем форму по центру
-		if(!attr.modal_dialog){
-			if((w = ((_cell.getWidth ? _cell.getWidth() : _cell.cell.offsetWidth) - 500)/2) >= 10)
-				_frm.cont.style.paddingLeft = w.toFixed() + "px";
-			else
-				_frm.cont.style.paddingLeft = "20px";
-		}
-
-		setTimeout(function () {
-
-			dhx4.callEvent("on_draw_auth", [_frm]);
-
-			if(($p.wsql.get_user_param("autologin") || attr.try_auto) && (selected || ($p.wsql.get_user_param("user_name") && $p.wsql.get_user_param("user_pwd"))))
-				auth_click.call(_frm);
-
-		});
-	});
-
-	// назначаем обработчик нажатия на кнопку
-	_frm.attachEvent("onButtonClick", auth_click);
-
-	_frm.attachEvent("onKeyDown",function(inp, ev, name, value){
-		if(ev.keyCode == 13){
-			if(name == "password" || this.getCheckedValue("type") == "guest"){
-				auth_click.call(this);
-			}
-		}
-	});
-
-
-	_frm.onerror = function (err) {
-
-		$p.ajax.authorized = false;
-
-		var emsg = err.message.toLowerCase();
-
-		if(emsg.indexOf("auth") != -1) {
-			$p.msg.show_msg({
-				title: $p.msg.main_title + $p.version,
-				type: "alert-error",
-				text: $p.msg.error_auth
-			});
-			_frm.setItemValue("password", "");
-			_frm.validate();
-
-		}else if(emsg.indexOf("gateway") != -1 || emsg.indexOf("net") != -1) {
-			$p.msg.show_msg({
-				title: $p.msg.main_title + $p.version,
-				type: "alert-error",
-				text: $p.msg.error_network
-			});
-		}
-	}
-
-
-
-};
-
-
-/**
- * Служебная функция для открытия окна настроек из гиперссылки
- * @param e
- * @return {Boolean}
- */
-$p.iface.open_settings = function (e) {
-	var evt = (e || (typeof event != "undefined" ? event : undefined));
-	if(evt)
-		evt.preventDefault();
-
-	var hprm = $p.job_prm.parse_url();
-	$p.iface.set_hash(hprm.obj, hprm.ref, hprm.frm, "settings");
-
-	return $p.iface.cancel_bubble(evt);
-};
-
-/**
- * Переключает вид формы между списком, календаарём и отчетами
- * @method swith_view
- * @for InterfaceObjs
- * @param name {String} - имя представления
- */
-$p.iface.swith_view = function(name){
-
-	var state,
-		iface = $p.iface,
-
-		/**
-		 * Переключает состав элементов дерева
-		 * @param view
-		 */
-		swith_tree = function(name){
-
-			function compare_text(a, b) {
-				if (a.text > b.text) return 1;
-				if (a.text < b.text) return -1;
-			}
-
-			if(!iface.tree){
-
-				var hprm = $p.job_prm.parse_url();
-				if(hprm.obj) {
-					var parts = hprm.obj.split('.');
-					if(parts.length > 1){
-
-						var mgr = $p.md.mgr_by_class_name(hprm.obj);
-
-						if(typeof iface.docs.close === "function" )
-							iface.docs.close();
-
-						if(mgr)
-							mgr.form_list(iface.docs, {});
-					}
-				}
-				return;
-
-			}else if(iface.tree._view == name || ["rep", "cal"].indexOf(name) != -1)
-				return;
-
-			iface.tree.deleteChildItems(0);
-			if(name == "oper"){
-				var meta_tree = {id:0, item:[
-					{id:"oper_cat", text: $p.msg.meta_cat, open: true, item:[]},
-					{id:"oper_doc", text: $p.msg.meta_doc, item:[]},
-					{id:"oper_cch", text: $p.msg.meta_cch, item:[]},
-					{id:"oper_cacc", text: $p.msg.meta_cacc, item:[]},
-					{id:"oper_tsk", text: $p.msg.meta_tsk, item:[]}
-				]}, mdn, md,
-
-				// бежим по справочникам
-					tlist = meta_tree.item[0].item;
-				for(mdn in $p.cat){
-					if(typeof $p.cat[mdn] == "function")
-						continue;
-					md = $p.cat[mdn].metadata();
-					if(md.hide)
-						continue;
-					tlist.push({id: "oper.cat." + mdn, text: md.synonym || md.name, tooltip: md.illustration || md.list_presentation});
-				}
-				tlist.sort(compare_text);
-
-				// бежим по документам
-				tlist = meta_tree.item[1].item;
-				for(mdn in $p.doc){
-					if(typeof $p.doc[mdn] == "function")
-						continue;
-					md = $p.doc[mdn].metadata();
-					if(md.hide)
-						continue;
-					tlist.push({id: "oper.doc." + mdn, text: md.synonym || md.name, tooltip: md.illustration || md.list_presentation});
-				}
-				tlist.sort(compare_text);
-
-				// бежим по планам видов характеристик
-				tlist = meta_tree.item[2].item;
-				for(mdn in $p.cch){
-					if(typeof $p.cch[mdn] == "function")
-						continue;
-					md = $p.cch[mdn].metadata();
-					if(md.hide)
-						continue;
-					tlist.push({id: "oper.cch." + mdn, text: md.synonym || md.name, tooltip: md.illustration || md.list_presentation});
-				}
-				tlist.sort(compare_text);
-
-				// бежим по планам счетов
-				tlist = meta_tree.item[3].item;
-				for(mdn in $p.cacc){
-					if(typeof $p.cacc[mdn] == "function")
-						continue;
-					md = $p.cacc[mdn].metadata();
-					if(md.hide)
-						continue;
-					tlist.push({id: "oper.cacc." + mdn, text: md.synonym || md.name, tooltip: md.illustration || md.list_presentation});
-				}
-				tlist.sort(compare_text);
-
-				// бежим по задачам
-				tlist = meta_tree.item[4].item;
-				for(mdn in $p.tsk){
-					if(typeof $p.tsk[mdn] == "function")
-						continue;
-					md = $p.tsk[mdn].metadata();
-					if(md.hide)
-						continue;
-					tlist.push({id: "oper.tsk." + mdn, text: md.synonym || md.name, tooltip: md.illustration || md.list_presentation});
-				}
-				tlist.sort(compare_text);
-
-				iface.tree.parse(meta_tree, function(){
-					var hprm = $p.job_prm.parse_url();
-					if(hprm.obj){
-						iface.tree.selectItem(hprm.view+"."+hprm.obj, true);
-					}
-				}, "json");
-
-			}else{
-				iface.tree.loadXML(iface.tree.tree_filteres, function(){
-
-				});
-
-			}
-
-			iface.tree._view = name;
-		};
-
-	if(name.indexOf(iface.docs.getViewName())==0)
-		return iface.docs.getViewName();
-
-	state = iface.docs.showView(name);
-	if (state == true) {
-		// first call, init corresponding components
-		// календарь
-		if(name=="cal" && !window.dhtmlXScheduler){
-			$p.load_script("dist/dhtmlxscheduler.min.js", "script", function(){
-				//scheduler.config.xml_date="%Y-%m-%d %H:%i";
-				scheduler.config.first_hour = 8;
-				scheduler.config.last_hour = 22;
-				iface.docs.scheduler = iface.docs.attachScheduler(new Date("2015-11-20"), "week", "scheduler_here");
-				iface.docs.scheduler.attachEvent("onBeforeViewChange", function(old_mode, old_date, mode, date){
-					if(mode == "timeline"){
-						$p.msg.show_not_implemented();
-						return false;
-					}
-					return true;
-				});
-			});
-
-			$p.load_script("dist/dhtmlxscheduler.css", "link");
-
-			//}else if(name=="rep"){
-			//	// подключаемый отчет
-			//
-			//}else if(name=="oper"){
-			//	// в дереве - список метаданных, в окне - список текущего метаданного
-			//
-
-		}
-	}
-
-	swith_tree(name);
-
-	if(name == "def")
-		iface.main.showStatusBar();
-	else
-		iface.main.hideStatusBar();
 };
 
 
@@ -746,11 +317,16 @@ function OTooolBar(attr){
 					this.subdiv = document.createElement('div');
 					this.subdiv.className = 'md_otooolbar';
 					offset = $p.iface.get_offset(bdiv);
-					if(battr.sub.align == 'right')
-						this.subdiv.style.left = (offset.left + bdiv.offsetWidth - (parseInt(battr.sub.width.replace(/\D+/g,"")) || 56)) + 'px';
-					else
-						this.subdiv.style.left = offset.left + 'px';
-					this.subdiv.style.top = (offset.top + div.offsetHeight) + 'px';
+					if(battr.sub.align == 'right') {
+            this.subdiv.style.left = (offset.left + bdiv.offsetWidth - (parseInt(battr.sub.width.replace(/\D+/g,"")) || 56)) + 'px';
+          }
+					else if(battr.sub.align == 'hor') {
+            this.subdiv.style.left = offset.left + bdiv.offsetWidth + 'px';
+          }
+					else{
+            this.subdiv.style.left = offset.left + 'px';
+          }
+					this.subdiv.style.top = (offset.top + (battr.sub.align == 'hor' ? 0 : div.offsetHeight)) + 'px';
 					this.subdiv.style.height = battr.sub.height || '198px';
 					this.subdiv.style.width = battr.sub.width || '56px';
 					for(var i in battr.sub.buttons){
@@ -863,8 +439,11 @@ $p.iface.add_button = function(parent, attr, battr) {
 		html +='<b style="vertical-align: super;"> ' + battr.b + '</b>';
 	else if(battr.text)
 		html +='<span style="vertical-align: super;"> ' + battr.text + '</span>';
-	else if(battr.css)
-		bdiv.classList.add(battr.css);
+	else if(battr.css){
+    battr.css.split(' ').forEach(function (s) {
+      s && bdiv.classList.add(s);
+    });
+  }
 	bdiv.innerHTML = html;
 
 	if(battr.float) bdiv.style.float = battr.float;
