@@ -23,7 +23,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	var _mgr = this,
 		_meta = _mgr.metadata(),
 		o = attr.o,
-		wnd, options, created, create_id, _title, close_confirmed;
+		wnd, options, created, create_id, _title;
 
 	/**
 	 * ПриСозданииНаСервере - инициализация при создании формы, до чтения объекта
@@ -57,7 +57,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 						// информируем мир о закрытии формы
 						if(_mgr && _mgr.class_name)
 							$p.eve.callEvent("frm_close", [_mgr.class_name, (o && o._obj ? o.ref : "")]);
-						
+
 						if(_wnd.conf){
 							_wnd.detachToolbar();
 							_wnd.detachStatusBar();
@@ -114,28 +114,17 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 						if(attr && attr.set_text || wnd && wnd.setText){
 
 							var title = o.presentation;
-							
+
 							if(!title)
 								return;
 
-							if(o instanceof CatObj)
-								title = (_meta.obj_presentation || _meta.synonym) + ': ' + title;
-
-							else if(o instanceof DocObj)
-								title += o.posted ? " (проведен)" : " (не проведен)";
-
-							if(o._modified && title.lastIndexOf("*")!=title.length-1)
-								title += " *";
-
-							else if(!o._modified && title.lastIndexOf("*")==title.length-1)
-								title = title.replace(" *", "");
+							if(o instanceof CatObj){
+                title = (_meta.obj_presentation || _meta.synonym) + ': ' + title;
+              }
 
 							if(force || _title !== title){
 								_title = title;
-								if(attr.set_text)
-									attr.set_text(title);
-								else
-									wnd.setText(title);
+								attr.set_text ? attr.set_text(title) : wnd.setText(title);
 							}
 						}
 					},
@@ -165,14 +154,14 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 		wnd.elmnts.frm_toolbar.loadStruct(attr.toolbar_struct || $p.injected_data["toolbar_obj.xml"], function(){
 
 			// если мы приклеены к ячейке, сдвигаем toolbar на 4px
-			if(wnd === pwnd)
-				this.cont.style.top = "4px";
+			// if(wnd === pwnd)
+			// 	this.cont.style.top = "4px";
 
 			this.addSpacer("btn_unpost");
 			this.attachEvent("onclick", attr.toolbar_click || toolbar_click);
 
 			// учтём права для каждой роли на каждый объект
-			var _acl = $p.current_acl.get_acl(_mgr.class_name);
+			var _acl = $p.current_user.get_acl(_mgr.class_name);
 
 			if(_mgr instanceof DocManager && _acl.indexOf("p") != -1){
 				this.enableItem("btn_post");
@@ -210,7 +199,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 					id: "btn_files"
 				});
 				wnd.elmnts.vault_pop.attachEvent("onShow", show_vault);
-				
+
 			}else
 				this.disableItem("bs_print");
 
@@ -230,9 +219,10 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	 * Пока здесь только установка заголовка формы
 	 * @param changes
 	 */
-	function observer(changes) {
-		if(o && wnd)
-			wnd.set_text();
+	function listener(obj, fields) {
+		if(wnd && wnd.set_text && ((obj === o) || (obj._owner && obj._owner._owner === o))){
+      wnd.set_text();
+    }
 	}
 
 	/**
@@ -249,8 +239,9 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 		 * Устанавливаем текст заголовка формы
 		 */
 		wnd.set_text();
-		if(!attr.hide_header && wnd.showHeader)
-			wnd.showHeader();
+		if(!attr.hide_header && wnd.showHeader){
+      wnd.showHeader();
+    }
 
 		/**
 		 * закладки табличных частей
@@ -285,7 +276,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 		else{
 
 			// учтём права для каждой роли на каждый объект
-			var _acl = $p.current_acl.get_acl(_mgr.class_name);
+			var _acl = $p.current_user.get_acl(_mgr.class_name);
 
 			wnd.elmnts.pg_header = wnd.elmnts.tabs.tab_header.attachHeadFields({
 				obj: o,
@@ -300,7 +291,10 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 		/**
 		 * начинаем следить за объектом
 		 */
-		Object.observe(o, observer, ["update", "row"]);
+    _mgr.on({
+      update: listener,
+      rows: listener,
+    });
 
 
 		return {wnd: wnd, o: o};
@@ -311,33 +305,33 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	 * обработчик нажатия кнопок командных панелей
 	 */
 	function toolbar_click(btn_id){
-		if(btn_id=="btn_save_close")
-			save("close");
-
-		else if(btn_id=="btn_save")
-			save("save");
-
-		else if(btn_id=="btn_post")
-			save("post");
-
-		else if(btn_id=="btn_unpost")
-			save("unpost");
-
-		else if(btn_id=="btn_close")
-			wnd.close();
-
-		else if(btn_id=="btn_go_connection")
-			go_connection();
-
-		else if(btn_id.substr(0,4)=="prn_")
-			_mgr.print(o, btn_id, wnd);
-
-		else if(btn_id=="btn_import")
-			_mgr.import(null, o);
-
-		else if(btn_id=="btn_export")
-			_mgr.export({items: [o], pwnd: wnd, obj: true} );
-
+		if(btn_id=="btn_save_close"){
+      save("close");
+    }
+    else if(btn_id=="btn_save"){
+      save("save");
+    }
+    else if(btn_id=="btn_post"){
+      save("post");
+    }
+    else if(btn_id=="btn_unpost"){
+      save("unpost");
+    }
+    else if(btn_id=="btn_close"){
+      wnd.close();
+    }
+    else if(btn_id=="btn_go_connection"){
+      go_connection();
+    }
+    else if(btn_id.substr(0,4)=="prn_"){
+      _mgr.print(o, btn_id, wnd);
+    }
+    else if(btn_id=="btn_import"){
+      _mgr.import(null, o);
+    }
+    else if(btn_id=="btn_export"){
+      _mgr.export({items: [o], pwnd: wnd, obj: true} );
+    }
 	}
 
 	/**
@@ -369,7 +363,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	/**
 	 * настройка (инициализация) табличной части
 	 */
-	function tabular_init(name, toolbar_struct){
+	function tabular_init(name, toolbar_struct, footer){
 
 		// с помощью метода ts_captions(), выясняем, надо ли добавлять данную ТЧ и формируем описание колонок табчасти
 		if(!_md.ts_captions(_mgr.class_name, name))
@@ -380,14 +374,15 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 		wnd.elmnts.tabs['tab_'+name] = wnd.elmnts.frm_tabs.cells('tab_'+name);
 
 		// учтём права для каждой роли на каждый объект
-		var _acl = $p.current_acl.get_acl(_mgr.class_name);
+		var _acl = $p.current_user.get_acl(_mgr.class_name);
 
 		wnd.elmnts.grids[name] = wnd.elmnts.tabs['tab_'+name].attachTabular({
 			obj: o,
 			ts: name,
 			pwnd: wnd,
 			read_only: _acl.indexOf("e") == -1,
-			toolbar_struct: toolbar_struct
+			toolbar_struct,
+      footer
 		});
 
 		if(_acl.indexOf("e") == -1){
@@ -405,7 +400,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	function save(action){
 
 		wnd.progressOn();
-		
+
 		var post;
 		if(o instanceof DocObj){
 			if(action == "post")
@@ -415,7 +410,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 				post = false;
 
 			else if(action == "close"){
-				if($p.current_acl.get_acl(_mgr.class_name).indexOf("p") != -1)
+				if($p.current_user.get_acl(_mgr.class_name).indexOf("p") != -1)
 					post = true;
 			}
 		}
@@ -429,7 +424,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 					if(attr.on_select)
 						attr.on_select(o);
 					wnd.close();
-					
+
 				}else
 					wnd.set_text();
 			})
@@ -445,13 +440,19 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	 */
 	function frm_unload(on_create){
 
-		if(attr && attr.on_close && !on_create)
-			attr.on_close();
+		if(attr && attr.on_close && !on_create){
+      attr.on_close();
+    }
 
 		if(!on_create){
-			delete wnd.ref;
-			delete wnd.set_text;
-			Object.unobserve(o, observer);
+		  if(_mgr){
+        _mgr.off('update', listener);
+        _mgr.off('rows', listener);
+      }
+      if(wnd){
+        delete wnd.ref;
+        delete wnd.set_text;
+      }
 			_mgr = wnd = o = _meta = options = pwnd = attr = null;
 		}
 	}
@@ -460,25 +461,26 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	 * Задаёт вопрос о записи изменений и делает откат при необходимости
 	 */
 	function check_modified() {
-		if(o._modified && !close_confirmed){
+		if(o && o._modified && !wnd.close_confirmed){
 			dhtmlx.confirm({
 				title: o.presentation,
 				text: $p.msg.modified_close,
 				cancel: $p.msg.cancel,
 				callback: function(btn) {
 					if(btn){
-						close_confirmed = true;
+            wnd.close_confirmed = true;
 						// закрыть изменённый без сохранения - значит прочитать его из pouchdb
-						if(o._manager.cachable == "ram")
-							this.close();
-
+						if(o._manager.cachable == "ram"){
+              this.close && this.close();
+            }
 						else{
 							if(o.is_new()){
 								o.unload();
-								this.close();
-							}else{
+                this.close && this.close();
+							}
+							else{
 								setTimeout(o.load.bind(o), 100);
-								this.close();
+                this.close && this.close();
 							}
 						}
 					}
@@ -492,7 +494,7 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	function frm_close(wnd){
 
 		if(check_modified()){
-			
+
 			setTimeout(frm_unload);
 
 			// выгружаем попапы
@@ -517,23 +519,22 @@ DataManager.prototype.form_obj = function(pwnd, attr){
 	// читаем объект из локального SQL или получаем с сервера
 	if($p.utils.is_data_obj(o)){
 
-		if(o.is_new() && attr.on_select)
-			return _mgr.create({}, true)
-				.then(function (tObj) {
-					o = tObj;
-					tObj = null;
-					return frm_fill();
-				});
+		if(o.is_new() && attr.on_select){
+      return _mgr.create({}, true)
+        .then(function (tObj) {
+          o = tObj;
+          tObj = null;
+          return frm_fill();
+        });
+    }
 		else if(o.is_new() && !o.empty()){
 			return o.load()
 				.then(frm_fill);
 		}else
 			return Promise.resolve(frm_fill());
-
-	}else{
-
-		if(pwnd && pwnd.progressOn)
-			pwnd.progressOn();
+	}
+	else{
+		pwnd && pwnd.progressOn && pwnd.progressOn();
 
 		return _mgr.get(attr.hasOwnProperty("ref") ? attr.ref : attr, true, true)
 			.then(function(tObj){
