@@ -22,23 +22,24 @@
  */
 DataManager.prototype.form_selection = function(pwnd, attr){
 
-	if(!pwnd)
-		pwnd = attr && attr.pwnd ? attr.pwnd : {};
+  if(!pwnd) {
+    pwnd = attr && attr.pwnd ? attr.pwnd : {};
+  }
 
-	if(!attr && !(pwnd instanceof dhtmlXCellObject)){
+  if(!attr && !(pwnd instanceof dhtmlXCellObject)){
 		attr = pwnd;
 		pwnd = {};
 	}
 
-	if(!attr)
-		attr = {};
+  if(!attr) {
+    attr = {};
+  }
 
 
-	var _mgr = this,
+  var _mgr = this,
 		_meta = attr.metadata || _mgr.metadata(),
 		has_tree = _meta["hierarchical"] && !(_mgr instanceof ChartOfAccountManager),
-		wnd, s_col = 0,
-		a_direction = "asc",
+		wnd, s_col = 0, a_direction = "asc",
 		previous_filter = {},
 		on_select = pwnd.on_select || attr.on_select;
 
@@ -73,7 +74,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 				});
 			}
 		}else{
-			wnd = $p.iface.w.createWindow(null, 0, 0, 700, 500);
+			wnd = $p.iface.w.createWindow(null, 0, 0, 760, 540);
 			wnd.centerOnScreen();
 			wnd.setModal(1);
 			wnd.button('park').hide();
@@ -83,16 +84,22 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 		}
 
 		$p.iface.bind_help(wnd);
+
 		if(wnd.setText && !attr.hide_text)
 			wnd.setText('Список ' + (_mgr.class_name.indexOf("doc.") == -1 ? 'справочника "' : 'документов "') + (_meta["list_presentation"] || _meta.synonym) + '"');
 
 		document.body.addEventListener("keydown", body_keydown, false);
 
 		// статусбар
-		wnd.elmnts = {
-			status_bar: wnd.attachStatusBar()
-		};
-		wnd.elmnts.status_bar.setText("<div id='" + _mgr.class_name.replace(".", "_") + "_select_recinfoArea'></div>");
+		wnd.elmnts = {}
+
+		if(attr.status_bar || !attr.smart_rendering){
+			wnd.elmnts.status_bar = wnd.attachStatusBar();
+		}
+
+		if(!attr.smart_rendering){
+			wnd.elmnts.status_bar.setText("<div id='" + _mgr.class_name.replace(".", "_") + "_select_recinfoArea'></div>");
+		}
 
 		// командная панель формы
 		wnd.elmnts.toolbar = wnd.attachToolbar();
@@ -105,7 +112,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 			if(wnd === pwnd){
 				this.cont.parentElement.classList.add("dhx_cell_toolbar_no_borders");
 				this.cont.parentElement.classList.remove("dhx_cell_toolbar_def");
-				this.cont.style.top = "4px";
+				//this.cont.style.top = "4px";
 			}
 
 			// текстовое поле фильтра по подстроке
@@ -121,9 +128,9 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 			if(attr.period) tbattr.period = attr.period;
 			wnd.elmnts.filter = new $p.iface.Toolbar_filter(tbattr);
 
-			
+
 			// учтём права для каждой роли на каждый объект
-			var _acl = $p.current_acl.get_acl(_mgr.class_name);
+			var _acl = $p.current_user.get_acl(_mgr.class_name);
 
 			if(_acl.indexOf("i") == -1)
 				this.hideItem("btn_new");
@@ -174,22 +181,9 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 	 */
 	function body_keydown(evt){
 
-		/**
-		 * Проверяет, нет ли других модальных форм
-		 */
-		function check_exit(){
-			var do_exit;
-			// если есть внешнее модальное, ничего обрабатывать не надо
-			$p.iface.w.forEachWindow(function (w) {
-				if(w != wnd && (w.isModal() || $p.iface.w.getTopmostWindow() == w))
-					do_exit = true;
-			});
-			return do_exit;
-		}
-
 		if(wnd && wnd.is_visible && wnd.is_visible()){
 			if (evt.ctrlKey && evt.keyCode == 70){ // фокус на поиск по {Ctrl+F}
-				if(!check_exit()){
+				if(!$p.iface.check_exit(wnd)){
 					setTimeout(function(){
 						if(wnd.elmnts.filter.input_filter && $p.job_prm.device_type == "desktop")
 							wnd.elmnts.filter.input_filter.focus();
@@ -198,7 +192,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 				}
 
 			} else if(evt.shiftKey && evt.keyCode == 116){ // requery по {Shift+F5}
-				if(!check_exit()){
+				if(!$p.iface.check_exit(wnd)){
 					setTimeout(function(){
 						wnd.elmnts.grid.reload();
 					});
@@ -208,7 +202,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 				}
 
 			} else if(evt.keyCode == 27){ // закрытие по {ESC}
-				if(!check_exit()){
+				if(wnd instanceof dhtmlXWindowsCell && !$p.iface.check_exit(wnd)){
 					setTimeout(function(){
 						wnd.close();
 					});
@@ -242,81 +236,131 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 			cell_tree.setWidth('220');
 			cell_tree.hideHeader();
 
-			tree = wnd.elmnts.tree = cell_tree.attachDynTree(_mgr, null, function(){
-				setTimeout(function(){
-					if(grid && grid.reload)
-						grid.reload();
-				}, 20);
-			});
-			tree.attachEvent("onSelect", function(id, mode){	// довешиваем обработчик на дерево
-				if(!mode)
-					return;
-				if(this.do_not_reload)
-					delete this.do_not_reload;
-				else
-					setTimeout(function(){
-						if(grid && grid.reload)
-							grid.reload();
-					}, 20);
-			});
-			tree.attachEvent("onDblClick", function(id){
+			const filter = {is_folder: true};
+      const {selection} = get_filter(0, 1000);
+      previous_filter = {};
+      if(Array.isArray(selection)) {
+        const set = new Set();
+        for (const sel of selection) {
+          for (let key in sel) {
+            if(key === 'ref') {
+              const cmp = sel[key].in ? 'in' : (sel[key].inh ? 'inh' : '')
+              if(cmp) {
+                sel[key][cmp].forEach((v) => {
+                  const o = _mgr.get(v);
+                  if(!o || o.empty()) {
+                    return;
+                  }
+                  o.is_folder && set.add(o);
+                  for (const elm of o._parents()) {
+                    set.add(elm);
+                  }
+                  for (const elm of o._children(true)) {
+                    set.add(elm);
+                  }
+                });
+              }
+            }
+          }
+        }
+        if(set.size) {
+          filter.ref = {in: Array.from(set)};
+        }
+      }
+      tree = wnd.elmnts.tree = cell_tree.attachDynTree(_mgr, filter, function(){
+        setTimeout(function () {
+          if(grid && grid.reload) {
+            grid.reload();
+          }
+        }, 20);
+      });
+      tree.attachEvent('onSelect', function (id, mode) {	// довешиваем обработчик на дерево
+        if(!mode) {
+          return;
+        }
+        if(this.do_not_reload) {
+          delete this.do_not_reload;
+        }
+        else {
+          setTimeout(function () {
+            if(grid && grid.reload) {
+              grid.reload();
+            }
+          }, 20);
+        }
+      });
+      tree.attachEvent("onDblClick", function(id){
 				select(id);
 			});
-
-		}else{
-			cell_grid = wnd;
-			setTimeout(function(){
-				if(grid && grid.reload)
-					grid.reload();
-			}, 20);
 		}
+		else{
+			cell_grid = wnd;
+      setTimeout(function () {
+        if(grid && grid.reload) {
+          grid.reload();
+        }
+      }, 20);
+    }
 
 		// настройка грида
 		grid = wnd.elmnts.grid = cell_grid.attachGrid();
 		grid.setIconsPath(dhtmlx.image_path);
 		grid.setImagePath(dhtmlx.image_path);
-		grid.setPagingWTMode(true,true,true,[20,30,60]);
-		grid.enablePaging(true, 30, 8, _mgr.class_name.replace(".", "_") + "_select_recinfoArea");
-		grid.setPagingSkin("toolbar", dhtmlx.skin);
 		grid.attachEvent("onBeforeSorting", customColumnSort);
 		grid.attachEvent("onBeforePageChanged", function(){ return !!this.getRowsNum();});
 		grid.attachEvent("onXLE", function(){cell_grid.progressOff(); });
 		grid.attachEvent("onXLS", function(){cell_grid.progressOn(); });
 		grid.attachEvent("onDynXLS", function(start,count){
-			var filter = get_filter(start,count);
-			if(!filter)
-				return;
-			_mgr.sync_grid(filter, grid);
-			return false;
-		});
+      var filter = get_filter(start, count);
+      if(!filter) {
+        return;
+      }
+      _mgr.sync_grid(filter, grid);
+      return false;
+    });
 		grid.attachEvent("onRowDblClicked", function(rId, cInd){
-			if(tree && tree.items[rId]){
-				tree.selectItem(rId);
-				var pid = tree.getParentId(rId);
-				if(pid && pid != $p.utils.blank.guid)
-					tree.openItem(pid);
-			}else
-				select(rId);
-		});
+      if(tree && tree.items[rId]) {
+        tree.selectItem(rId);
+        var pid = tree.getParentId(rId);
+        if(pid && pid != $p.utils.blank.guid) {
+          tree.openItem(pid);
+        }
+      }
+      else {
+        select(rId);
+      }
+    });
 
-		if($p.iface.docs && $p.iface.docs.getViewName && $p.iface.docs.getViewName() == "oper")
-			grid.enableMultiselect(true);
+    if(attr.smart_rendering) {
+      grid.enableSmartRendering(true, 50);
+    }
+    else {
+      grid.setPagingWTMode(true, true, true, [20, 30, 60]);
+      grid.enablePaging(true, 30, 8, _mgr.class_name.replace('.', '_') + '_select_recinfoArea');
+      grid.setPagingSkin('toolbar', dhtmlx.skin);
+    }
 
-		// эту функцию будем вызывать снаружи, чтобы перечитать данные
+    if($p.iface.docs && $p.iface.docs.getViewName && $p.iface.docs.getViewName() == 'oper') {
+      grid.enableMultiselect(true);
+    }
+
+    // эту функцию будем вызывать снаружи, чтобы перечитать данные
 		grid.reload = function(){
 
 			var filter = get_filter();
-			if(!filter)
-				return Promise.resolve();
+      if(!filter) {
+        return Promise.resolve();
+      }
 
-			cell_grid.progressOn();
+      cell_grid.progressOn();
 			grid.clearAll();
 
 			return _mgr.sync_grid(filter, grid)
 				.then(function(xml){
 					if(typeof xml === "object"){
 						$p.msg.check_soap_result(xml);
-					}else if(!grid_inited){
+					}
+					else if(!grid_inited){
 						if(filter.initial_value){
 							var xpos = xml.indexOf("set_parent"),
 								xpos2 = xml.indexOf("'>", xpos),
@@ -337,17 +381,20 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 						grid.enableAutoWidth(true, 1200, 600);
 						grid.setSizes();
 						grid_inited = true;
-						if(wnd.elmnts.filter.input_filter && $p.job_prm.device_type == "desktop")
-							wnd.elmnts.filter.input_filter.focus();
+            if(wnd.elmnts.filter.input_filter && $p.job_prm.device_type == 'desktop') {
+              wnd.elmnts.filter.input_filter.focus();
+            }
 
-						if(attr.on_grid_inited)
-							attr.on_grid_inited();
-					}
+            if(attr.on_grid_inited) {
+              attr.on_grid_inited();
+            }
+          }
 
-					if (a_direction && grid_inited)
-						grid.setSortImgState(true, s_col, a_direction);
+          if(a_direction && grid_inited) {
+            grid.setSortImgState(true, s_col, a_direction);
+          }
 
-					cell_grid.progressOff();
+          cell_grid.progressOff();
 
 				});
 		};
@@ -358,65 +405,69 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 	 */
 	function toolbar_click(btn_id){
 
+		// если внешний обработчик вернул false - выходим
+		if(attr.toolbar_click && attr.toolbar_click(btn_id, wnd, _mgr) === false){
+			return;
+		}
+
 		if(btn_id=="btn_select"){
 			select();
-
-		}else if(btn_id=="btn_new"){
-
+		}
+		else if(btn_id=="btn_new"){
 			_mgr.create({}, true)
 				.then(function (o) {
-
-					if(attr.on_new)
-						attr.on_new(o, wnd);
-
+					if(attr.on_new){
+            attr.on_new(o, wnd);
+          }
 					else if($p.job_prm.keep_hash){
 						o.form_obj(wnd);
-
-					} else{
+					}
+					else{
 						o._set_loaded(o.ref);
 						$p.iface.set_hash(_mgr.class_name, o.ref);
 					}
 				});
-
-
-		}else if(btn_id=="btn_edit") {
-			var rId = wnd.elmnts.grid.getSelectedRowId();
+		}
+		else if(btn_id=="btn_edit") {
+			const rId = wnd.elmnts.grid.getSelectedRowId();
 			if (rId){
-				if(attr.on_edit)
-					attr.on_edit(_mgr, rId, wnd);
-
-				else if($p.job_prm.keep_hash){
-
+				if(attr.on_edit){
+          attr.on_edit(_mgr, rId, wnd);
+        }
+        else if($p.job_prm.keep_hash){
 					_mgr.form_obj(wnd, {ref: rId});
 
-				} else
-					$p.iface.set_hash(_mgr.class_name, rId);
-			}else
-				$p.msg.show_msg({
-					type: "alert-warning",
-					text: $p.msg.no_selected_row.replace("%1", ""),
-					title: $p.msg.main_title
-				});
-
-		}else if(btn_id.substr(0,4)=="prn_"){
+				}
+				else{
+          $p.iface.set_hash(_mgr.class_name, rId);
+        }
+			}
+			else{
+        $p.msg.show_msg({
+          type: "alert-warning",
+          text: $p.msg.no_selected_row.replace("%1", ""),
+          title: $p.msg.main_title
+        });
+      }
+		}
+		else if(btn_id.substr(0,4)=="prn_"){
 				print(btn_id);
-
-		}else if(btn_id=="btn_order_list"){
+		}
+		else if(btn_id=="btn_order_list"){
 			$p.iface.set_hash("", "", "", "def");
-
-		}else if(btn_id=="btn_delete"){
+		}
+		else if(btn_id=="btn_delete"){
 			mark_deleted();
-
-		}else if(btn_id=="btn_import"){
+		}
+		else if(btn_id=="btn_import"){
 			_mgr.import();
-
-		}else if(btn_id=="btn_export"){
+		}
+		else if(btn_id=="btn_export"){
 			_mgr.export(wnd.elmnts.grid.getSelectedRowId());
-
-		}else if(btn_id=="btn_requery"){
+		}
+		else if(btn_id=="btn_requery"){
 			previous_filter = {};
 			wnd.elmnts.grid.reload();
-
 		}
 	}
 
@@ -470,7 +521,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 
 			else if(on_select){
 
-				_mgr.get(rId, true)
+				_mgr.get(rId, 'promise')
 					.then(function(selv){
 						wnd.close();
 						on_select.call(pwnd.grid || pwnd, selv);
@@ -502,7 +553,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 	function mark_deleted(){
 		var rId = wnd.elmnts.grid.getSelectedRowId();
 		if(rId){
-			_mgr.get(rId, true, true)
+			_mgr.get(rId, 'promise')
 				.then(function (o) {
 
 					dhtmlx.confirm({
@@ -528,8 +579,9 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 
 		document.body.removeEventListener("keydown", body_keydown);
 
-		if(attr && attr.on_close && !on_create)
-			attr.on_close();
+		if(attr && attr.on_close && !on_create){
+      attr.on_close();
+    }
 
 		if(!on_create){
 			_mgr = wnd = _meta = previous_filter = on_select = pwnd = attr = null;
@@ -559,32 +611,41 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 	 * @return {*|{value, enumerable}}
 	 */
 	function get_filter(start, count){
-		var filter = wnd.elmnts.filter.get_filter()
+	  const {grid, tree} = wnd.elmnts;
+    const filter = wnd.elmnts.filter.get_filter()
 				._mixin({
 					action: "get_selection",
 					metadata: _meta,
 					class_name: _mgr.class_name,
-					order_by: wnd.elmnts.grid.columnIds[s_col] || s_col,
+					order_by: (grid && grid.columnIds[s_col]) || s_col,
 					direction: a_direction,
-					start: start || ((wnd.elmnts.grid.currentPage || 1)-1)*wnd.elmnts.grid.rowsBufferOutSize,
-					count: count || wnd.elmnts.grid.rowsBufferOutSize,
+          start: start || (grid ? ((grid.currentPage || 1) - 1) * grid.rowsBufferOutSize : 0),
+					count: count || (grid ? grid.rowsBufferOutSize : 50),
 					get_header: (previous_filter.get_header == undefined)
 				}),
-			tparent = has_tree ? wnd.elmnts.tree.getSelectedId() : null;
+			tparent = (has_tree && tree) ? tree.getSelectedId() : null;
 
-		if(attr.date_from && !filter.date_from)
-			filter.date_from = attr.date_from;
+    if(attr.smart_rendering) {
+      filter.smart_rendering = true;
+    }
 
-		if(attr.date_till && !filter.date_till)
-			filter.date_till = attr.date_till;
+    if(attr.date_from && !filter.date_from) {
+      filter.date_from = attr.date_from;
+    }
 
-		if(attr.initial_value)
-			filter.initial_value = attr.initial_value;
+    if(attr.date_till && !filter.date_till) {
+      filter.date_till = attr.date_till;
+    }
 
-		if(attr.custom_selection)
-			filter.custom_selection = attr.custom_selection;
+    if(attr.initial_value) {
+      filter.initial_value = attr.initial_value;
+    }
 
-		if(attr.selection){
+    if(attr.custom_selection) {
+      filter.custom_selection = attr.custom_selection;
+    }
+
+    if(attr.selection){
 			if(!filter.selection)
 				filter.selection = attr.selection;
 
@@ -607,15 +668,17 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 			//}
 		}
 
-		if(attr.owner && !filter.owner)
-			filter.owner = attr.owner;
+    if(attr.owner && !filter.owner) {
+      filter.owner = attr.owner;
+    }
 
-		filter.parent = ((tparent  || attr.parent) && !filter.filter) ? (tparent || attr.parent) : null;
-		if(has_tree && !filter.parent)
-			filter.parent = $p.utils.blank.guid;
+    filter.parent = ((tparent  || attr.parent) && !filter.filter) ? (tparent || attr.parent) : null;
+    if(has_tree && !filter.parent) {
+      filter.parent = $p.utils.blank.guid;
+    }
 
 
-		for(var f in filter){
+    for(var f in filter){
 			if(previous_filter[f] != filter[f]){
 				previous_filter = filter;
 				return filter;
@@ -630,9 +693,9 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 		wnd.elmnts.grid.reload();
 		return true;
 	}
-	
+
 	/**
-	 * подписываемся на событие закрытия формы объекта, чтобы обновить список и попытаться спозиционироваться на нужной строке 
+	 * подписываемся на событие закрытия формы объекта, чтобы обновить список и попытаться спозиционироваться на нужной строке
 	 */
 	var _frm_close = $p.eve.attachEvent("frm_close", function (class_name, ref) {
 		if(_mgr && _mgr.class_name == class_name && wnd && wnd.elmnts){
@@ -646,7 +709,7 @@ DataManager.prototype.form_selection = function(pwnd, attr){
 
 	// создаём и настраиваем форму
 	if(has_tree && attr.initial_value && attr.initial_value!= $p.utils.blank.guid && !attr.parent)
-		return _mgr.get(attr.initial_value, true)
+		return _mgr.get(attr.initial_value, 'promise')
 			.then(function (tObj) {
 				attr.parent = tObj.parent.ref;
 				attr.set_parent = attr.parent;
